% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventTable.R
\name{add.dosing}
\alias{add.dosing}
\title{Add dosing to eventTable}
\usage{
add.dosing(eventTable, ...)
}
\arguments{
\item{eventTable}{eventTable object}

\item{...}{arguments sent to eventTable$add.dosing.}
}
\value{
eventTable with updated dosing (note the event table will be updated anyway)
}
\description{
This adds a dosing event to the event table.  This is provided for
piping syntax through magrittr
}
\seealso{
\code{\link{eventTable}}, \code{\link{RxODE}}
}
\author{
Matthew L. Fidler
}
