% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventTable.R
\name{add.sampling}
\alias{add.sampling}
\title{Add sampling to eventTable}
\usage{
add.sampling(eventTable, time, time.units = NA, ...)
}
\arguments{
\item{eventTable}{An eventTable object}

\item{time}{a vector of time values (in \code{time.units}).}

\item{time.units}{an optional string specifying the time
units. Defaults to the units specified when the
\code{EventTable} was initialized.}

\item{...}{Other parameters (ignored)}
}
\value{
eventTable with updated sampling.  (Note the event table
    will be updated even if you don't reassign the eventTable)
}
\description{
This adds a dosing event to the event table.  This is provided for
piping syntax through magrittr
}
\seealso{
\code{\link{eventTable}}, \code{\link{RxODE}}
}
\author{
Matthew L. Fidler
}
