% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{arrange_.solveRxDll}
\alias{arrange_.solveRxDll}
\title{dyplr arrange_ support}
\usage{
arrange_.solveRxDll(.data, ..., .dots)

arrange_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that arrange_ is being applied to.}

\item{...}{arguments to arrange_ method}

\item{.dots}{dplyr .dots argument}

\item{.data}{solveRxDll object that arrange_ is being applied to.}

\item{...}{arguments to arrange_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr arrange_ support

dyplr arrange_ support
}
\keyword{internal}

