% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{as.data.frame.rxEvid}
\alias{as.data.frame.rxEvid}
\title{Convert to data.frame}
\usage{
\method{as.data.frame}{rxEvid}(x, row.names = NULL, optional = FALSE,
  ..., nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse =
  " "))
}
\arguments{
\item{x}{any \R object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{additional arguments to be passed to or from methods.}

\item{nm}{Name of column in new data frame}
}
\description{
Convert to data.frame
}
