% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{as.data.frame.solveRxDll}
\alias{as.data.frame.solveRxDll}
\title{Convert solveRxDll object to data.frame}
\usage{
\method{as.data.frame}{solveRxDll}(x, row.names = NULL, optional = FALSE,
  ..., stringsAsFactors = default.stringsAsFactors())
}
\description{
Extract the matrix returned by the solver and then convert it to a
data.frame
}

