% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{as.data.table.rxEt}
\alias{as.data.table.rxEt}
\title{Convert an event table to a data.table}
\usage{
as.data.table.rxEt(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{An R object.}

\item{keep.rownames}{Default is \code{FALSE}. If \code{TRUE}, adds the input object's names as a separate column named \code{"rn"}. \code{keep.rownames = "id"} names the column \code{"id"} instead.}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\description{
Convert an event table to a data.table
}
