% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{as.data.table.solveRxDll}
\alias{as.data.table.solveRxDll}
\title{as.data.table for solveRxDll}
\usage{
\method{as.data.table}{solveRxDll}(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{Solved ODE, a \code{solveRxDll} object}

\item{keep.rownames}{Keep rownames when changing to data tables (default FALSE)}

\item{...}{Other parameters for as.data.table}
}
\description{
compatability function for data.table
}
