% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{as.et}
\alias{as.et}
\alias{as.et.default}
\title{Coerce object to data.frame}
\usage{
as.et(x, ...)

\method{as.et}{default}(x, ...)
}
\arguments{
\item{x}{Object to coerce to et.}

\item{...}{Other parameters}
}
\description{
Coerce object to data.frame
}
