% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{as.matrix.solveRxDll}
\alias{as.matrix.solveRxDll}
\title{Convert solveRxDll object to matrix}
\usage{
\method{as.matrix}{solveRxDll}(x, ...)
}
\description{
This is done by extracting the matrix from the solveRxDll object.
}

