% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{as.tbl.solveRxDll}
\alias{arrange_.solveRxDll}
\alias{as.tbl.solveRxDll}
\alias{distinct_.solveRxDll}
\alias{filter_.solveRxDll}
\alias{gather_.solveRxDll}
\alias{group_by_.solveRxDll}
\alias{mutate_.solveRxDll}
\alias{rename_.solveRxDll}
\alias{sample_frac.solveRxDll}
\alias{sample_n.solveRxDll}
\alias{select_.solveRxDll}
\alias{separate_.solveRxDll}
\alias{slice_.solveRxDll}
\alias{spread_.solveRxDll}
\alias{summarise_.solveRxDll}
\alias{transmute_.solveRxDll}
\alias{unite_.solveRxDll}
\title{Compatability functions for solveRxDll with dyplr and tidyr}
\usage{
as.tbl.solveRxDll(x, ...)

filter_.solveRxDll(.data, ..., .dots)

slice_.solveRxDll(.data, ..., .dots)

arrange_.solveRxDll(.data, ..., .dots)

select_.solveRxDll(.data, ..., .dots)

rename_.solveRxDll(.data, ..., .dots)

distinct_.solveRxDll(.data, ..., .dots)

mutate_.solveRxDll(.data, ..., .dots)

transmute_.solveRxDll(.data, ..., .dots)

summarise_.solveRxDll(.data, ..., .dots)

arrange_.solveRxDll(.data, ..., .dots)

rename_.solveRxDll(.data, ..., .dots)

group_by_.solveRxDll(.data, ..., .dots, add = FALSE)

sample_n.solveRxDll(tbl, size, replace = FALSE, weight = NULL,
  .env = parent.frame())

sample_frac.solveRxDll(tbl, size = 1, replace = FALSE, weight = NULL,
  .env = parent.frame())

gather_.solveRxDll(data, key_col, value_col, gather_cols, na.rm = FALSE,
  convert = FALSE, factor_key = FALSE)

separate_.solveRxDll(data, col, into, sep = "[^[:alnum:]]+", remove = TRUE,
  convert = FALSE, extra = "warn", fill = "warn", ...)

unite_.solveRxDll(data, col, from, sep = "_", remove = TRUE)

spread_.solveRxDll(data, key_col, value_col, fill = NA, convert = FALSE,
  drop = TRUE)
}
\description{
Compatability functions for solveRxDll with dyplr and tidyr
}

