% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{as_data_frame.solveRxDll}
\alias{as_data_frame.solveRxDll}
\title{Use the as_data_frame method for solved object x}
\usage{
as_data_frame.solveRxDll(x)
}
\arguments{
\item{x}{Solved RxODE object}
}
\value{
data frame of solved object.
}
\description{
Use the as_data_frame method for solved object x
}
\author{
Matthew L. Fidler
}
