% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{as_tibble.rxEt}
\alias{as_tibble.rxEt}
\title{Convert to tbl}
\usage{
\method{as_tibble}{rxEt}(x, ...)
}
\arguments{
\item{x}{RxODE event table}

\item{...}{Other arguments to \code{as_tibble}}
}
\value{
tibble of event table
}
\description{
Convert to tbl
}
