% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{$.solveRxDll}
\alias{$.solveRxDll}
\title{Add data-frame like operators to solved objects}
\usage{
\method{$}{solveRxDll}(obj, arg)
}
\description{
This is a little different from a standard data frame accessor,
because it allows access to other model parameters for the solved object.
}

