% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{$<-.solveRxDll}
\alias{$<-.solveRxDll}
\title{$ Assign for RxODE solved objects}
\usage{
\method{$}{solveRxDll}(obj, arg) <- value
}
\arguments{
\item{obj}{solveRxDll object}

\item{arg}{Dollar sign name}

\item{value}{assigned Value.}
}
\description{
Assign objects by argumnt as obj$arg <- value
}
\details{
This also works as obj[[arg]] <- value
}
\seealso{
\code{\link{rxSolve}}
}
\keyword{internal}

