% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{rbind.solveRxDll}
\alias{rbind.solveRxDll}
\alias{cbind.solveRxDll}
\title{cbind/rbind solveRxDll}
\usage{
\method{rbind}{solveRxDll}(...)

\method{cbind}{solveRxDll}(...)
}
\arguments{
\item{...}{ignored parameters}
}
\description{
Cbind/rbind is disabled for RxOde solved objects.  Use as.data.frame(x)
to use cbind/rbind.
}
\author{
Matthew L.Fidler

Matthew L.Fidler
}
\keyword{internal}
