% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{dim<-.solveRxDll}
\alias{dim<-.solveRxDll}
\title{Assign dimensions of rxSolve object}
\usage{
\method{dim}{solveRxDll}(x) <- value
}
\arguments{
\item{x}{rxSolve object}

\item{value}{dimensions}
}
\description{
dim(x) <- value
}
\keyword{internal}

