% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{dim.solveRxDll}
\alias{dim.solveRxDll}
\title{Get dimensions of rxSolve object}
\usage{
\method{dim}{solveRxDll}(x)
}
\arguments{
\item{x}{rxSolve object}
}
\description{
Get dimensions of rxSolve object
}
\keyword{internal}

