% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{dimnames<-.solveRxDll}
\alias{dimnames<-.solveRxDll}
\title{Assign dimension names for rxSolve object}
\usage{
\method{dimnames}{solveRxDll}(x) <- value
}
\arguments{
\item{x}{rxSolve object}

\item{value}{dimension names assigned}
}
\description{
dimnames(x) <- value
}
\keyword{internal}

