% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{dimnames.solveRxDll}
\alias{dimnames.solveRxDll}
\title{Get dimension names for rxSolve object}
\usage{
\method{dimnames}{solveRxDll}(x)
}
\arguments{
\item{x}{rxSolve object}
}
\description{
Get dimension names for rxSolve object
}
\keyword{internal}

