% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{distinct_}
\alias{distinct_}
\alias{distinct_.solveRxDll}
\title{distinct_ for \code{solveRxDll} object}
\usage{
\method{distinct_}{solveRxDll}(.data, ...)
}
\arguments{
\item{.data}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
compatability function for dplyr
}
