% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{distinct_.solveRxDll}
\alias{distinct_.solveRxDll}
\title{dyplr distinct_ support}
\usage{
distinct_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that distinct_ is being applied to.}

\item{...}{arguments to distinct_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr distinct_ support
}
\keyword{internal}

