% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxFinalizePred}
\alias{.rxFinalizePred}
\title{Finalize RxODE pred based on symengine saved info}
\usage{
.rxFinalizePred(.s, sum.prod = FALSE, optExpression = TRUE)
}
\arguments{
\item{.s}{Symengine/RxODE object}

\item{sum.prod}{A boolean determining if RxODE should use more
numerically stable sums/products.}

\item{optExpression}{Optimize the model text for computer
evaluation.}
}
\value{
Nothing
}
\description{
Finalize RxODE pred based on symengine saved info
}
\author{
Matthew L Fidler
}
