% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxOptExpr.R
\name{.rxFindPow}
\alias{.rxFindPow}
\title{Find power THETAs for appropriate scaling}
\usage{
.rxFindPow(x)
}
\arguments{
\item{x}{RxODE model that can be access by rxNorm}
}
\value{
THETA numbers of x^theta
}
\description{
Find power THETAs for appropriate scaling
}
\author{
Matthew L. Fidler
}
