% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenEBE}
\alias{.rxGenEBE}
\title{Generate peices for EBE only problem}
\usage{
.rxGenEBE(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL,
  pred.minus.dv = TRUE, sum.prod = FALSE, optExpression = TRUE)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{pred.minus.dv}{Boolean stating if the FOCEi objective
function is based on PRED-DV (like NONMEM).  Default TRUE.}

\item{sum.prod}{A boolean determining if RxODE should use more
numerically stable sums/products.}

\item{optExpression}{Optimize the model text for computer
evaluation.}
}
\value{
RxODE/symengine environment
}
\description{
Generate peices for EBE only problem
}
\author{
Matthew L. Fidler
}
