% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenEtaS}
\alias{.rxGenEtaS}
\title{Generate the ETA sensitivities for FO related methods}
\usage{
.rxGenEtaS(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL,
  promoteLinSens = FALSE)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}
}
\value{
RxODE/symengine environment
}
\description{
Generate the ETA sensitivities for FO related methods
}
\author{
Matthew L. Fidler
}
