% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenFocei}
\alias{.rxGenFocei}
\title{Generate peices for FOCEi inner problem}
\usage{
.rxGenFocei(
  obj,
  predfn,
  pkpars = NULL,
  errfn = NULL,
  init = NULL,
  pred.minus.dv = TRUE,
  sum.prod = FALSE,
  optExpression = TRUE,
  promoteLinSens = TRUE,
  theta = FALSE
)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{pred.minus.dv}{Boolean stating if the FOCEi objective
function is based on PRED-DV (like NONMEM).  Default TRUE.}

\item{sum.prod}{A boolean determining if RxODE should use more
numerically stable sums/products.}

\item{optExpression}{Optimize the model text for computer
evaluation.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}

\item{theta}{Calculate THETA derivatives instead of ETA
derivatives.  By default FALSE}
}
\value{
RxODE/symengine environment
}
\description{
Generate peices for FOCEi inner problem
}
\author{
Matthew L. Fidler
}
