% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenFun}
\alias{.rxGenFun}
\title{Generate and load RxODE function into sympy environment}
\usage{
.rxGenFun(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL)
}
\value{
Sympy environment
}
\description{
Generate and load RxODE function into sympy environment
}
\author{
Matthew Fidler
}
