% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenFunState}
\alias{.rxGenFunState}
\title{Get the state information for the current model}
\usage{
.rxGenFunState(obj)
}
\arguments{
\item{obj}{RxODE model that can take rxModelVars}
}
\value{
character vector of initial preserved states (state),
extra states (statef) and dvid translation information
(dvid). This is used in generating the final RxODE model.
}
\description{
Get the state information for the current model
}
\author{
Matthew Fidler
}
