% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenHdEta}
\alias{.rxGenHdEta}
\title{Generate the d(err)/d(eta) values for FO related methods}
\usage{
.rxGenHdEta(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL,
  pred.minus.dv = TRUE, promoteLinSens = TRUE, theta = FALSE)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{pred.minus.dv}{Boolean stating if the FOCEi objective
function is based on PRED-DV (like NONMEM).  Default TRUE.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}

\item{theta}{Calculate THETA derivatives instead of ETA
derivatives.  By default FALSE}
}
\value{
RxODE/symengine environment
}
\description{
Generate the d(err)/d(eta) values for FO related methods
}
\author{
Matthew L. Fidler
}
