% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenPred}
\alias{.rxGenPred}
\title{Generate Augmented Pred/err RxODE model}
\usage{
.rxGenPred(obj, predfn, errfn, init)
}
\value{
RxODE model variables augmented with pred/error information
}
\description{
Generate Augmented Pred/err RxODE model
}
\author{
Matthew L Fidler
}
