% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err-foceiBase.R
\name{.rxGetVarianceForErrorAddPow}
\alias{.rxGetVarianceForErrorAddPow}
\title{Additive + Power}
\usage{
.rxGetVarianceForErrorAddPow(env, pred1)
}
\arguments{
\item{env}{Environment for the parsed model}

\item{pred1}{The \code{data.frame} of the current error}
}
\value{
additive + power
}
\description{
Additive + Power
}
\author{
Matthew Fidler
}
