% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R, R/rxsympy.R
\name{.rxRmIni}
\alias{.rxRmIni}
\title{Remove INIs}
\usage{
.rxRmIni(x)

.rxRmIni(x)
}
\arguments{
\item{x}{RxODE list of lines to remove}

\item{x}{RxODE list of lines to remove}
}
\value{
RxODE lines with inis removed.

RxODE lines with inis removed.
}
\description{
Remove INIs

Remove INIs
}
\author{
Matthew L. Fidler

Matthew L. Fidler
}
\keyword{internal}
