% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{.rxRmJac}
\alias{.rxRmJac}
\title{Remove Jacobian}
\usage{
.rxRmJac(x)
}
\arguments{
\item{x}{RxODE list of lines to remove}
}
\value{
RxODE lines with \code{df/dy} removed.
}
\description{
Remove Jacobian
}
\author{
Matthew L. Fidler
}
\keyword{internal}
