% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R, R/rxsympy.R
\name{.rxRmPrint}
\alias{.rxRmPrint}
\title{Remove print statements}
\usage{
.rxRmPrint(x)

.rxRmPrint(x)
}
\arguments{
\item{x}{RxODE lines to remove}

\item{x}{RxODE lines to remove}
}
\value{
RxODE with print lines removed.

RxODE with print lines removed.
}
\description{
Remove print statements

Remove print statements
}
\author{
Matthew L. Fidler

Matthew L. Fidler
}
