% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{.rxRmSens}
\alias{.rxRmSens}
\title{Remove sensitivity equations}
\usage{
.rxRmSens(x)
}
\arguments{
\item{x}{RxODE lines to remove}
}
\value{
Lines with d/dt(rx_sens_...._) removed.
}
\description{
Remove sensitivity equations
}
\author{
Matthew L. Fidler
}
