% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE-win-setup.R
\name{.rxRtoolsBaseWin}
\alias{.rxRtoolsBaseWin}
\title{Return Rtools base}
\usage{
.rxRtoolsBaseWin()
}
\value{
Rtools base path, or "" on unix-style platforms.
}
\description{
Return Rtools base
}
\author{
Matthew L. Fidler
}
