% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE-win-setup.R
\name{.rxWinRtoolsPath}
\alias{.rxWinRtoolsPath}
\title{Setup Rtools path}
\usage{
.rxWinRtoolsPath(rm.rtools = TRUE, retry = FALSE)
}
\arguments{
\item{rm.rtools}{Remove the Rtools from the current path specs.}

\item{retry}{Should you retry to find Rtools?  If NA, don't throw
an error if it isn't found.}
}
\description{
Setup Rtools path
}
\author{
Matthew L. Fidler
}
