% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rxWithOptions}
\alias{.rxWithOptions}
\title{Temporarily set options then restore them while running code}
\usage{
.rxWithOptions(ops, code)
}
\arguments{
\item{ops}{Options to temporarily set}

\item{code}{The code to run during the sink}
}
\description{
Temporarily set options then restore them while running code
}
\examples{

.rxWithOptions(digits=21, {
   print(pi)
})
}
