% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rxWithWd}
\alias{.rxWithWd}
\title{Temporarily set options then restore them while running code}
\usage{
.rxWithWd(wd, code)
}
\arguments{
\item{wd}{working directory to temporarily set the system to while
evaluating the code}

\item{code}{The code to run during the sink}
}
\value{
value of code
}
\description{
Temporarily set options then restore them while running code
}
\examples{

.rxWithWd(tempdir(), {
   getwd()
})

getwd()
}
