% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etTran.R
\name{.setWarnIdSort}
\alias{.setWarnIdSort}
\title{Turn on/off warnings for ID sorting.}
\usage{
.setWarnIdSort(warnIdSort = TRUE)
}
\arguments{
\item{warnIdSort}{Boolean for if the sorting warning is turned on or off.}
}
\value{
Nothing
}
\description{
Turn on/off warnings for ID sorting.
}
\author{
Matthew Fidler
}
