% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{edit.solveRxDll}
\alias{edit.solveRxDll}
\title{Edit data frame of rxSolve}
\usage{
\method{edit}{solveRxDll}(name, ...)
}
\arguments{
\item{name}{name of the object}
}
\description{
Edit data frame of rxSolve
}
\keyword{internal}

