% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etExpand}
\alias{etExpand}
\title{Expand additional doses}
\usage{
etExpand(et)
}
\arguments{
\item{et}{Event table to expand additional doses for.}
}
\value{
New event table with \code{addl} doses expanded
}
\description{
Expand additional doses
}
\examples{
ev <- et(amt = 3, ii = 24, until = 240)
print(ev)
etExpand(ev) # expands event table, but doesn't modify it

print(ev)

ev$expand() ## Expands the current event table and saves it in ev
}
\author{
Matthew Fidler
}
