% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etRbind}
\alias{etRbind}
\title{Combining event tables}
\usage{
etRbind(..., samples = c("use", "clear"), waitII = c("smart", "+ii"),
  id = c("merge", "unique"))
}
\arguments{
\item{...}{The event tables and optionally time between event
tables, called waiting times in this help document.}

\item{samples}{How to handle samples when repeating an event
table.  The options are: \itemize{ \item{"clear"} Clear
sampling records before combining the datasets \item{"use"}
Use the sampling records when combining the datasets }}

\item{waitII}{This determines how waiting times between events are
    handled. The options are:

\itemize{

\item \code{"smart"} This "smart" handling of waiting times is the
default option.  In this case, if the waiting time is above the
last observed inter-dose interval in the first combined event
table, then the actual time between doses is given by the wait
time.  If it is smaller than the last observed inter-dose
interval, the time between event tables is given by the inter-dose
interval + the waiting time between event tables.

\item \code{"+ii"} In this case, the wait time is added to the
inter-dose interval no matter the length of the wait time or
inter-dose interval

}}
}
\value{
A new event table
}
\description{
Combining event tables
}
