% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etRep}
\alias{etRep}
\title{Repeat an RxODE event table}
\usage{
etRep(x, times = 1, length.out = NA, each = NA, n = NULL,
  wait = 0, id = integer(0), samples = c("clear", "use"),
  waitII = c("smart", "+ii"), ii = 24)
}
\arguments{
\item{times}{Number of times to repeat the event table}

\item{length.out}{Invalid with RxODE event tables, will throw an
error if used.}

\item{each}{Invalid with RxODE event tables, will throw an error
if used.}

\item{n}{The number of times to repeat the event table.  Overrides
\code{times}.}

\item{wait}{Waiting time between each repeated event table.  By
default there is no waiting, or wait=0}

\item{id}{IDs to expand/remove in the event table before repeating.}

\item{samples}{How to handle samples when repeating an event
table.  The options are: \itemize{ \item{"clear"} Clear
sampling records before combining the datasets \item{"use"}
Use the sampling records when combining the datasets }}

\item{waitII}{This determines how waiting times between events are
    handled. The options are:

\itemize{

\item \code{"smart"} This "smart" handling of waiting times is the
default option.  In this case, if the waiting time is above the
last observed inter-dose interval in the first combined event
table, then the actual time between doses is given by the wait
time.  If it is smaller than the last observed inter-dose
interval, the time between event tables is given by the inter-dose
interval + the waiting time between event tables.

\item \code{"+ii"} In this case, the wait time is added to the
inter-dose interval no matter the length of the wait time or
inter-dose interval

}}

\item{ii}{If there was no inter-dose intervals found in the event
table, assume that the interdose interval is given by this
\code{ii} value.  By default this is \code{24}.}
}
\value{
An event table of repeated events
}
\description{
Repeat an RxODE event table
}
\author{
Matthew Fidler
}
