% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etSeq}
\alias{etSeq}
\title{Sequence of event tables}
\usage{
etSeq(..., samples = c("clear", "use"), waitII = c("smart", "+ii"),
  ii = 24)
}
\arguments{
\item{samples}{can be "clear", "use"}
}
\value{
A new event table
}
\description{
Sequence of event tables
}
\details{

}
\author{
Matthew L Fidler
}
