% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{filter_}
\alias{filter_}
\alias{filter_.solveRxDll}
\title{filter_ for \code{solveRxDll} object}
\usage{
\method{filter_}{solveRxDll}(.data, ...)
}
\arguments{
\item{.data}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
compatability function for dplyr
}
