% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{filter_.solveRxDll}
\alias{filter_.solveRxDll}
\title{dyplr filter_ support}
\usage{
filter_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that filter_ is being applied to.}

\item{...}{arguments to filter_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr filter_ support
}
\keyword{internal}

