% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei2.R
\name{foceiFit}
\alias{foceiFit}
\title{FOCEi fit}
\usage{
foceiFit(data, inits, PKpars, model = NULL, pred = NULL, err = NULL,
  lower = -Inf, upper = Inf, fixed = NULL, skipCov = NULL,
  control = foceiControl(), thetaNames = NULL, etaNames = NULL, ...)
}
\arguments{
\item{data}{Data to fit}

\item{inits}{Initialization list}

\item{PKpars}{Pk Parameters}

\item{model}{The RxODE model to use}

\item{pred}{The Prediction function}

\item{err}{The Error function}

\item{lower}{Lower bounds}

\item{upper}{Upper Bounds}

\item{fixed}{Boolean vector indicating what parameters should be fixed.}

\item{skipCov}{Boolean vector indicating what parameters should be fixed when calculating covariances}

\item{control}{FOCEi options Control list}

\item{thetaNames}{Names of the thetas to be used in the final object}

\item{etaNames}{Eta names to be used in the final object}

\item{...}{Ignored parameters}
}
\value{
A focei fit object

FOCEi fit object
}
\description{
FOCEi fit
}
\author{
Matthew L. Fidler and Wenping Wang

Matthew L. Fidler & Wenping Wang
}
