% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forderForceBase}
\alias{forderForceBase}
\title{Force using base order for RxODE radix sorting}
\usage{
forderForceBase(forceBase = FALSE)
}
\arguments{
\item{forceBase}{boolean indicating if RxODE should use R's
\code{\link[=order]{order()}} for radix sorting instead of
\code{data.table}'s parallel radix sorting.}
}
\value{
NILL; called for side effects
}
\description{
Force using base order for RxODE radix sorting
}
\examples{
\donttest{
forderForceBase(TRUE) # Use base `order` for RxODE sorts
forderForceBase(FALSE) # Use `data.table` for RxODE sorts
}
}
