% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gammap}
\alias{gammap}
\title{Gammap: normalized lower incomplete gamma function}
\usage{
gammap(a, z)
}
\arguments{
\item{a}{The numeric 'a' parameter in the normalized lower
incomplete gamma}

\item{z}{The numeric 'z' parameter in the normalized lower
incomplete gamma}
}
\value{
gammap results
}
\description{
This is the gamma_p from the boost library
}
\details{
The gamma p function is given by:

gammap = lowergamma(a, z)/gamma(a)
}
\examples{

gammap(1,3)
gammap(1:3,3)
gammap(1,1:3)

}
\author{
Matthew L. Fidler
}
