% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gammapDer}
\alias{gammapDer}
\title{gammapDer:  derivative of gammap}
\usage{
gammapDer(a, z)
}
\arguments{
\item{a}{The numeric 'a' parameter in the upper
incomplete gamma}

\item{z}{The numeric 'z' parameter in the upper
incomplete gamma}
}
\value{
lowergamma results
}
\description{
This is the gamma_p_derivative from the boost library
}
\examples{

gammapDer(1:3,3)

gammapDer(1,1:3)

}
\author{
Matthew L. Fidler
}
