% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gammaq}
\alias{gammaq}
\title{Gammaq: normalized upper incomplete gamma function}
\usage{
gammaq(a, z)
}
\arguments{
\item{a}{The numeric 'a' parameter in the normalized upper
incomplete gamma}

\item{z}{The numeric 'z' parameter in the normalized upper
incomplete gamma}
}
\value{
gammaq results
}
\description{
This is the gamma_q from the boost library
}
\details{
The gamma q function is given by:

gammaq = uppergamma(a, z)/gamma(a)
}
\examples{

gammaq(1, 3)
gammaq(1:3, 3)
gammaq(1, 1:3)
}
\author{
Matthew L. Fidler
}
