% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{gather_}
\alias{gather_}
\alias{gather_.solveRxDll}
\title{gather_ for \code{solveRxDll} object}
\usage{
\method{gather_}{solveRxDll}(data, ...)
}
\arguments{
\item{data}{Solved ODE, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
compatability function for tidyr
}
