% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRxThreads}
\alias{getRxThreads}
\alias{setRxThreads}
\alias{rxCores}
\title{Get/Set the number of threads that RxODE uses}
\usage{
getRxThreads(verbose = FALSE)

setRxThreads(threads = NULL, percent = NULL, throttle = NULL)

rxCores(verbose = FALSE)
}
\arguments{
\item{verbose}{Display the value of relevant OpenMP settings}

\item{threads}{NULL (default) rereads environment variables. 0
means to use all logical CPUs available. Otherwise a number >= 1}

\item{percent}{If provided it should be a number between 2 and
100; the percentage of logical CPUs to use. By default on
startup, 50 percent.}

\item{throttle}{1024 (default) means that, roughly speaking, a
single thread will be used when number of all points<=1024, 2 threads when
the number of all points is <=2048, etc. The throttle is to speed up small data
tasks (especially when repeated many times) by not incurring the
overhead of managing multiple threads. Hence the number of
threads is throttled (restricted) for small tasks.}
}
\description{
Get/Set the number of threads that RxODE uses
}
