% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{group_by_}
\alias{group_by_}
\alias{group_by_.solveRxDll}
\title{group_by_ for \code{solveRxDll} object}
\usage{
\method{group_by_}{solveRxDll}(.data, ...)
}
\arguments{
\item{.data}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
dplyr compatability layer for solveRxDll
}
