% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{group_by_.solveRxDll}
\alias{group_by_.solveRxDll}
\title{dyplr group_by_ support}
\usage{
group_by_.solveRxDll(.data, ..., .dots, add = FALSE)
}
\arguments{
\item{.data}{solveRxDll object that groupt_by_ is being applied to.}

\item{...}{arguments to groupt_by_ method}

\item{.dots}{dplyr .dots argument}

\item{add}{dpylr \code{add} argument.  By default, when \code{add =
FALSE}, \code{group_by} will override existing groups.  To
instead add to the existing groups, use \code{add = TRUE}.}
}
\description{
dyplr group_by_ support
}
\keyword{internal}

