% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{guide_none}
\alias{guide_none}
\title{Empty Guide}
\usage{
guide_none(title = waiver(), position = waiver())
}
\description{
This empty guide draws nothing; It is included in RxODE for
compatibility with ggplot 3.2
}
