% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{head.solveRxDll}
\alias{head.solveRxDll}
\title{Return the head of RxODE solved data}
\usage{
\method{head}{solveRxDll}(x, n = 6L, ...)
}
\description{
Extract the matrix and return the head of the matrix
}

