% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{idrInfo}
\alias{idrInfo}
\title{IDR information}
\usage{
idrInfo(cmt = "eff", text = "+Kin-Kout*(1-centr/V2/(EC50+centr/V2))*eff")
}
\arguments{
\item{cmt}{Compartment that idr information is being assesed on.}

\item{text}{Text that is parsed for indirect response information.}
}
\value{
IDR information (if any)
}
\description{
IDR information parsing routine
}
\author{
Matthew L. Fidler
}
\keyword{internal}
