% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{is.rxEt}
\alias{is.rxEt}
\title{Check to see if this is an rxEt object.}
\usage{
is.rxEt(x)
}
\arguments{
\item{x}{object to check to see if it is rxEt

If this is an rxEt object that has expired strip all rxEt
information.}
}
\description{
Check to see if this is an rxEt object.
}
\author{
Matthew L.Fidler
}
