% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{is.rxSolve}
\alias{is.rxSolve}
\title{Check to see if this is an rxSolve object.}
\usage{
is.rxSolve(x)
}
\arguments{
\item{x}{object to check to see if it is rxSolve

If this is an rxSolve object that has expired strip all rxSolve
information.}
}
\value{
boolean indicating if this is a \code{rxSolve} object
}
\description{
Check to see if this is an rxSolve object.
}
\author{
Matthew L.Fidler
}
\keyword{internal}
