% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\alias{expit}
\title{logit and inverse logit (expit) functions}
\usage{
logit(x, low = 0, high = 1)

expit(alpha, low = 0, high = 1)
}
\arguments{
\item{x}{Input value(s) in range [low,high] to translate -Inf to
Inf}

\item{low}{Lowest value in the range}

\item{high}{Highest value in the range}

\item{alpha}{Infinite value(s) to translate to range of [low,
high]}
}
\description{
logit and inverse logit (expit) functions
}
\details{
logit is given by:

\deqn{logit(p) = -log(1/p-1)}

where:

\deqn{p = \frac{x-low}{high-low}}

 expit is given by:

\deqn{expit(p, low, high) = \frac{high-low}{1+exp(-alpha)}+low}
}
\examples{

logit(0.25)

expit()

}
