% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lowergamma}
\alias{lowergamma}
\title{lowergamma:  upper incomplete gamma function}
\usage{
lowergamma(a, z)
}
\arguments{
\item{a}{The numeric 'a' parameter in the upper
incomplete gamma}

\item{z}{The numeric 'z' parameter in the upper
incomplete gamma}
}
\value{
lowergamma results
}
\description{
This is the tgamma_lower from the boost library
}
\details{
The lowergamma function is given by:

\deqn{lowergamma(a, z) = \int_{0}^{z}t^{a-1}\cdot e^{-t} dt}
}
\examples{

lowergamma(1,3)

lowergamma(1:3,3)

lowergamma(1,1:3)

}
\author{
Matthew L. Fidler
}
