% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{model}
\alias{model}
\alias{model.default}
\title{Model block for RxODE/nlmixr models}
\usage{
model(x, ..., envir = parent.frame())

\method{model}{default}(x, ...)
}
\arguments{
\item{x}{model expression}

\item{...}{Other arguments}
}
\value{
Model block with ini information included.  \code{ini} must be called before \code{model} block
}
\description{
Model block for RxODE/nlmixr models
}
\author{
Matthew Fidler
}
