% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{mutate_.solveRxDll}
\alias{mutate_.solveRxDll}
\title{dyplr mutate_ support}
\usage{
mutate_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that mutate_ is being applied to.}

\item{...}{arguments to mutate_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr mutate_ support
}
\keyword{internal}

