% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\docType{data}
\name{oral1cmt}
\alias{oral1cmt}
\title{RxODE one compartment model (solved)}
\format{An object of class \code{RxODE} of length 37.}
\usage{
oral1cmt
}
\description{
RxODE one compartment model (solved)
}
\examples{
oral1cmt \%>\% solve(et(timeUnits="hr") \%>\% et(amt=100)) \%>\% plot
}
\keyword{datasets}
