% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\docType{data}
\name{oral2cmt}
\alias{oral2cmt}
\title{RxODE two compartment model (solved)}
\format{An object of class \code{RxODE} of length 37.}
\usage{
oral2cmt
}
\description{
RxODE two compartment model (solved)
}
\examples{
oral2cmt \%>\% solve(et(timeUnits="hr") \%>\% et(amt=100)) \%>\% plot
}
\keyword{datasets}
