% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\docType{data}
\name{oral3cmt}
\alias{oral3cmt}
\title{RxODE three compartment model (solved)}
\format{An object of class \code{RxODE} of length 37.}
\usage{
oral3cmt
}
\description{
RxODE three compartment model (solved)
}
\examples{
oral3cmt \%>\% solve(et(timeUnits="hr") \%>\% et(amt=100)) \%>\% plot
}
\keyword{datasets}
