% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phi}
\alias{phi}
\title{Cumulative distribution of standard normal}
\usage{
phi(q)
}
\arguments{
\item{q}{vector of quantiles.}
}
\value{
cumulative distribution of standard normal distribution
}
\description{
Cumulative distribution of standard normal
}
\author{
Matthew Fidler
}
