% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{pillar_shaft}
\alias{pillar_shaft}
\title{Re export of pillar_shaft}
\usage{
pillar_shaft(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Unused, for extensibility.}
}
\description{
Re export of pillar_shaft
}
