% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{pillar_shaft.rxRateDur}
\alias{pillar_shaft.rxRateDur}
\title{Pillar shaft for rxRateDur}
\usage{
\method{pillar_shaft}{rxRateDur}(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Unused, for extensibility.}
}
\description{
Pillar shaft for rxRateDur
}
