% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{plot.RxODE}
\alias{plot.RxODE}
\alias{plot.rxDll}
\alias{plot.RxCompilationManager}
\title{Plot a model digram for simple ODE models}
\usage{
\method{plot}{RxODE}(x, ...)

\method{plot}{rxDll}(x, ...)

\method{plot}{RxCompilationManager}(x, ...)
}
\arguments{
\item{x}{is an RxODE object.}

\item{...}{other arguments, see \code{\link{rxPlot}} for full
description.}
}
\description{
Plot a model digram for simple ODE models
}
\seealso{
\code{\link{rxPlot}},\code{\link{RxODE}}.
}
