% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{+.solveRxODE}
\alias{+.solveRxODE}
\title{Update Solved object with '+'}
\usage{
\method{+}{solveRxODE}(solved, new)
}
\arguments{
\item{solved}{Solved object}

\item{new}{New information added tothe table.}
}
\value{
new solved object
}
\description{
Update Solved object with '+'
}
\author{
Matthew L. Fidler
}
\keyword{internal}
