% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{predict.RxODE}
\alias{predict.RxODE}
\title{Predict an RxODE object}
\usage{
\method{predict}{RxODE}(object, ...)
}
\arguments{
\item{object}{An RxODE object}

\item{...}{Solve arguments sent to \code{rxSolve}.  See
\code{\link{rxSolve}}.}
}
\description{
\code{predict} solves the odinary differential equations specified by
 RxODE object
}

