% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rxDll}
\alias{print.rxDll}
\title{Print rxDll object}
\usage{
\method{print}{rxDll}(x, ...)
}
\value{
original object
}
\description{
This tells if the rxDll is loaded, ready and/or deleted.
}
\author{
Matthew L.Fidler
}
\keyword{internal}
