% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{print.RxODE}
\alias{print.RxODE}
\title{Print information about the RxODE object.}
\usage{
\method{print}{RxODE}(RxODEobj, ...)
}
\arguments{
\item{RxODEobj}{An rxode object}

\item{...}{Ignored parameters}
}
\description{
This prints the model name and its status for being able to be solved
}

