% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{print.solveRxDll}
\alias{print.solveRxDll}
\title{Print information about solved object}
\usage{
\method{print}{solveRxDll}(x, ...)
}
\arguments{
\item{x}{An solveRxDll object}

\item{...}{Ignored Parameter}
}
\description{
This prints the data frame and parameters of the model being
solved.
}

