% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probit}
\alias{probit}
\alias{probitInv}
\title{probit and inverse probit functions}
\usage{
probit(x, low = 0, high = 1)

probitInv(x, low = 0, high = 1)
}
\arguments{
\item{x}{Input value(s) in range [low,high] to translate -Inf to
Inf}

\item{low}{Lowest value in the range}

\item{high}{Highest value in the range}
}
\value{
values from probit, probitInv and probitNormInfo
}
\description{
probit and inverse probit functions
}
\examples{

probit(0.25)

probitInv(-0.674)

probitNormInfo(probit(0.25),sd=0.1)

probitNormInfo(probit(1,0,10), sd=1, low=0, high=10)

}
