% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{facet_wrap_paginate}
\alias{facet_grid_paginate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggforce}{\code{\link[ggforce]{facet_wrap_paginate}}, \code{\link[ggforce]{facet_grid_paginate}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

