% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{rename_}
\alias{rename_}
\alias{rename_.solveRxDll}
\alias{rename_}
\alias{rename_.solveRxDll}
\title{rename_ for \code{solveRxDll} object}
\usage{
\method{rename_}{solveRxDll}(.data, ...)

\method{rename_}{solveRxDll}(.data, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{data}{Solved equation, an \code{solveRxDll} object.}

\item{data}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
