% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rename_.solveRxDll}
\alias{rename_.solveRxDll}
\title{dyplr rename_ support}
\usage{
rename_.solveRxDll(.data, ..., .dots)

rename_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that rename_ is being applied to.}

\item{...}{arguments to rename_ method}

\item{.dots}{dplyr .dots argument}

\item{.data}{solveRxDll object that rename_ is being applied to.}

\item{...}{arguments to rename_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr rename_ support

dyplr rename_ support
}
\keyword{internal}

