% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{riwish}
\alias{riwish}
\title{Simulate one deviate from the inverse Wishart distribution}
\usage{
riwish(nu, Omega)
}
\arguments{
\item{nu}{Degrees of freedom of inverse Wishart distribution.}

\item{Omega}{Positive definite Omega Scale index.}
}
\value{
One random deviate (matrix) of inverse Wishart Distribution.
}
\description{
Simulate one deviate from the inverse Wishart distribution
}
\author{
Matthew Fidler
}
