% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{riwishDf}
\alias{riwishDf}
\title{Simulate one deviate from the inverse Wishart distribution multiplied by nu}
\usage{
riwishDf(nu, Omega)
}
\arguments{
\item{nu}{Degrees of freedom of inverse Wishart distribution.}

\item{Omega}{Positive definite Omega Scale index.}
}
\value{
One random deviate (matrix) of inverse Wishart Distribution scaled by the degrees of freedom.
}
\description{
Simulate one deviate from the inverse Wishart distribution multiplied by nu
}
\author{
Matthew Fidler & Wenping Wang
}
