% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rLKJ1}
\alias{rLKJ1}
\title{One correlation sample from the LKJ distribution}
\usage{
rLKJ1(d, eta = 1, cholesky = FALSE)
}
\arguments{
\item{d}{The dimension of the correlation matrix}

\item{eta}{The scaling parameter of the LKJ distribution.
Must be > 1.  Also related to the degrees of freedom nu.
eta = (nu-1)/2.}

\item{cholesky}{boolean; If \code{TRUE} return the cholesky
decomposition.}
}
\value{
A correlation sample from the LKJ distribution
}
\description{
One correlation sample from the LKJ distribution
}
\author{
Matthew Fidler (translated to RcppArmadillo) and Emma Schwager
}
