% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{row.names<-.solveRxDll}
\alias{row.names<-.solveRxDll}
\title{Assign rownames to rxSolve object}
\usage{
\method{row.names}{solveRxDll}(x) <- value
}
\arguments{
\item{x}{rxode object}

\item{value}{value assigned.}
}
\description{
row.names(x) <- value;
}
\keyword{internal}

