% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{row.names.solveRxDll}
\alias{row.names.solveRxDll}
\title{Get row names of rxSolve object}
\usage{
\method{row.names}{solveRxDll}(x, ...)
}
\arguments{
\item{x}{rxSovle object}

\item{...}{ignored arguments}
}
\description{
Get row names of rxSolve object
}
\keyword{internal}

