% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsiwish}
\alias{rsiwish}
\title{Simulate one deviate from the scaled inverse Wishart distribution}
\usage{
rsiwish(nu, Omega, mu, delta)
}
\arguments{
\item{nu}{Degrees of freedom of inverse Wishart distribution.}

\item{Omega}{Positive definite Omega Scale index.}

\item{mu}{vector of location hyper-parameters.}

\item{delta}{vector of location scale hyper-parameters.}
}
\value{
One random deviate (matrix) of inverse Wishart Distribution.
}
\description{
Simulate one deviate from the scaled inverse Wishart distribution
}
\author{
Matthew Fidler
}
