% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwish}
\alias{rwish}
\title{Simulate one deviate from the Wishart distribution}
\usage{
rwish(nu, Omega)
}
\arguments{
\item{nu}{Degrees of freedom of Wishart distribution.}

\item{Omega}{Positive definite Omega Scale matrix.}
}
\value{
One random deviate (matrix) of Wishart Distribution.
}
\description{
Simulate one deviate from the Wishart distribution
}
\author{
Matthew Fidler
}
