% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rx.do.call}
\alias{rx.do.call}
\title{do.call while setting up windows path for compiling}
\usage{
rx.do.call(..., envir = parent.frame())
}
\arguments{
\item{...}{Arguments sent to do.call}

\item{envir}{Environment to do the call in}
}
\value{
do.call return
}
\description{
do.call while setting up windows path for compiling
}
\author{
Matthew L. Fidler
}
\keyword{internal}
