% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R, R/rxsympy.R
\name{rxAddReturn}
\alias{rxAddReturn}
\title{Add a return statment to a function.}
\usage{
rxAddReturn(fn, ret = TRUE)

rxAddReturn(fn, ret = TRUE)
}
\arguments{
\item{fn}{Function to deparse}

\item{ret}{boolean stating if a return statement will be added.}

\item{fn}{Function to deparse}

\item{ret}{boolean stating if a return statement will be added.}
}
\value{
Function with parens removed and add a return statment.

Function with parens removed and add a return statement.
}
\description{
Add a return statment to a function.

Add a return statement to a function.
}
\author{
Matthew L. Fidler

Matthew L. Fidler
}
