% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxAllowUnload}
\alias{rxAllowUnload}
\title{Allow unloading of dlls}
\usage{
rxAllowUnload(allow)
}
\arguments{
\item{allow}{boolean indicating if garbage collection will unload of RxODE dlls.}
}
\description{
Allow unloading of dlls
}
\examples{

# Garbage collection will not unload un-used RxODE dlls
rxAllowUnload(FALSE);

# Garbage collection will unload unused RxODE dlls
rxAllowUnload(TRUE);
}
\author{
Matthew Fidler
}
