% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxAssignPtr}
\alias{rxAssignPtr}
\title{Assign pointer based on model variables}
\usage{
rxAssignPtr(object = NULL)
}
\arguments{
\item{object}{RxODE family of objects}
}
\value{
nothing, called for side effects
}
\description{
Assign pointer based on model variables
}
