% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxSymInv.R
\name{rxBlockZeros}
\alias{rxBlockZeros}
\title{Creates a logical matrix for block matrixes.}
\usage{
rxBlockZeros(mat, i)
}
\arguments{
\item{mat}{Matrix}

\item{i}{Row/column where block matrix should be setup.}
}
\value{
A logical matrix returning where the elements should be
    zero.
}
\description{
Creates a logical matrix for block matrixes.
}
\keyword{internal}
