% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxC}
\alias{rxC}
\title{Return the C file associated with the RxODE object}
\usage{
rxC(obj)
}
\arguments{
\item{obj}{A RxODE family of objects or a character string of the
model specification or location of a file with a model
specification.}
}
\value{
a path of the library
}
\description{
This will return C code for generating the RxODE DLL.
}
\author{
Matthew L.Fidler
}
\keyword{internal}
