% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxCat}
\alias{rxCat}
\title{Use cat when RxODE.verbose is TRUE}
\usage{
rxCat(a, ...)
}
\arguments{
\item{...}{Parameters sent to cat}
}
\description{
Use cat when RxODE.verbose is TRUE
}
\author{
Matthew L. Fidler
}
\keyword{internal}
