% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\docType{data}
\name{rxChain}
\alias{+.solveRxDll}
\alias{\%>\%.solveRxDll}
\alias{rxChain}
\title{Add item to solved system of equations}
\format{An object of class \code{character} of length 1.}
\usage{
\method{+}{solveRxDll}(obj1, obj2)

\%>\%.solveRxDll
}
\arguments{
\item{obj1}{Solved object}

\item{obj2}{New object to be added/piped/chained to solved object.

When \code{newObject} is an event table, return a new solved object
with the new event table.}
}
\description{
Add item to solved system of equations
}
\keyword{datasets}

