% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxClean}
\alias{rxClean}
\title{Cleanup anonymous dlls}
\usage{
rxClean(wd = getwd())
}
\arguments{
\item{wd}{What directory should be cleand

This cleans up all files named rx-*.dll and associated files as
well as call_dvode.o and associated files}
}
\value{
TRUE if successful
#'
}
\description{
This cleans up any dlls created by text files
}

