% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err-sim.R
\name{rxCombineErrorLines}
\alias{rxCombineErrorLines}
\title{Combine Error Lines and create RxODE expression}
\usage{
rxCombineErrorLines(uiModel, errLines = NULL)
}
\arguments{
\item{uiModel}{UI model}

\item{errLines}{Error lines; If missing, get the error lines from
\code{rxGetDistributionSimulationLines()}}
}
\value{
quoted extression that can be evaluated to compiled RxODE
model
}
\description{
Combine Error Lines and create RxODE expression
}
\examples{


one.cmt <- function() {
   ini({
     ## You may label each parameter with a comment
     tka <- 0.45 # Log Ka
     tcl <- log(c(0, 2.7, 100)) # Log Cl
     ## This works with interactive models
     ## You may also label the preceding line with label("label text")
     tv <- 3.45; label("log V")
     ## the label("Label name") works with all models
     eta.ka ~ 0.6
     eta.cl ~ 0.3
     eta.v ~ 0.1
     add.sd <- 0.7
   })
   model({
     ka <- exp(tka + eta.ka)
     cl <- exp(tcl + eta.cl)
      v <- exp(tv + eta.v)
      linCmt() ~ add(add.sd)
   })
}

f <- RxODE(one.cmt)

# You can get the simulation model easily by
rxCombineErrorLines(f)

# You can then get the compiled model by simply evaluting the model:
r <- eval(rxCombineErrorLines(f))

}
\author{
Matthew L. Fidler
}
\keyword{internal}
