% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxCondition}
\alias{rxCondition}
\title{Current Condition for RxODE object}
\usage{
rxCondition(obj, condition = NULL)
}
\arguments{
\item{obj}{RxODE object}

\item{condition}{If specified and is one of the conditions in the
RxODE object (as determined by \code{\link[=rxExpandIfElse]{rxExpandIfElse()}}),
assign the RxODE current condition to this parameter.  If the
condition is not one of the known condition, the condition is
set to \code{NULL}, implying no conditioning currently used.}
}
\value{
Current condition for RxODE object
}
\description{
Current Condition for RxODE object
}
\author{
Matthew L. Fidler
}
\keyword{internal}
