% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxCout}
\alias{rxCout}
\title{Give the cout string for a number}
\usage{
rxCout(number)
}
\arguments{
\item{number}{Number for converting}
}
\value{
String of cout << number;
}
\description{
Give the cout string for a number
}
\author{
Matthew L. Fidler
}
\keyword{internal}
