% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxCoutEcho}
\alias{rxCoutEcho}
\title{Echo cout to console for a number}
\usage{
rxCoutEcho(number)
}
\arguments{
\item{number}{number to output

 @return nothing.}
}
\description{
Echo cout to console for a number
}
