% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode-options.R
\name{rxCreateCache}
\alias{rxCreateCache}
\title{This will create the cache directory for RxODE to save between sessions}
\usage{
rxCreateCache()
}
\value{
nothing
}
\description{
When run, if the \code{R_user_dir} for RxODE's cache isn't present,
create the cache
}
\author{
Matthew Fidler
}
