% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxD}
\alias{rxD}
\title{Add to RxODE's derivative tables}
\usage{
rxD(name, derivatives)
}
\arguments{
\item{name}{Function Name}

\item{derivatives}{A list of functions. Each function takes the
same number of arguments as the original function.  The first
function will construct the derivative with respect to the
first argument; The second function will construct the
derivitive with respect to the second argument, and so on.}
}
\value{
nothing
}
\description{
Add to RxODE's derivative tables
}
\examples{
## Add an arbitrary list of derivative functions
## In this case the fun(x,y) is assumed to be 0.5*x^2+0.5*y^2

rxD("fun", list(
  function(x, y) {
    return(x)
  },
  function(x, y) {
    return(y)
  }
))
}
\author{
Matthew Fidler
}
