% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxData}
\alias{rxData}
\title{Setup a data frame for solving multiple subjects at once in RxODE.}
\usage{
rxData(object, params = NULL, events = NULL, inits = NULL, covs = NULL,
  method = "lsoda", transit_abs = NULL, atol = 1e-08, rtol = 1e-06,
  maxsteps = 5000L, hmin = 0L, hmax = NULL, hini = 0L, maxordn = 12L,
  maxords = 5L, cores = 1L, covs_interpolation = "linear",
  addCov = FALSE, matrix = FALSE, sigma = NULL, sigmaDf = NULL,
  sigmaNcores = 1L, sigmaIsChol = FALSE, amountUnits = NA_character_,
  timeUnits = "hours")
}
\arguments{
\item{object}{R object to setup; Must be in RxODE compatible format.}

\item{params}{a numeric named vector with values for every
parameter in the ODE system; the names must correspond to the
parameter identifiers used in the ODE specification;}

\item{events}{an \code{eventTable} object describing the input
(e.g., doses) to the dynamic system and observation sampling
time points (see \code{\link{eventTable}});}

\item{inits}{a vector of initial values of the state variables
(e.g., amounts in each compartment), and the order in this
vector must be the same as the state variables (e.g., PK/PD
compartments);}

\item{covs}{a matrix or dataframe the same number of rows as the
sampling points defined in the events \code{eventTable}.  This
is for time-varying covariates.}

\item{transit_abs}{boolean indicating if this is a transit
compartment absorption}

\item{atol}{a numeric absolute tolerance (1e-08 by default) used
by the ODE solver to determine if a good solution has been
achieved;}

\item{rtol}{a numeric relative tolerance (1e-06 by default) used
by the ODE solver to determine if a good solution has been
achieved.}

\item{maxsteps}{maximum number of (internally defined) steps allowed
during one call to the solver. (5000 by default)}

\item{hmin}{The minimum absolute step size allowed. The default
value is 0.}

\item{hmax}{The maximum absolute step size allowed.  The default
checks for the maximum difference in times in your sampling and
events, and uses this value.  The value 0 is equivalent to
infinite maximum absolute step size.}

\item{hini}{The step size to be attempted on the first step. The
default value is determined by the solver (when hini = 0)}

\item{maxordn}{The maximum order to be allowed for the nonstiff
(Adams) method.  The default is 12.  It can be between 1 and
12.}

\item{maxords}{The maximum order to be allowed for the stiff (BDF)
method.  The default value is 5.  This can be between 1 and 5.}

\item{covs_interpolation}{specifies the interpolation method for
    time-varying covariates. When solving ODEs it often samples
    times outside the sampling time specified in \code{events}.
    When this happens, the time varying covariates are
    interpolated.  Currently this can be:

\itemize{
\item \code{"linear"} interpolation (the default), which interpolates the covariate
    by solving the line between the observed covariates and extrapolating the new
    covariate value.
\item \code{"constant"} -- Last observation carried forward.
\item \code{"NOCB"} -- Next Observation Carried Backward.  This is the same method
      that NONMEM uses.
\item \code{"midpoint"} Last observation carried forward to midpoint; Next observation
  carried backward to midpoint.
}}

\item{matrix}{A boolean inticating if a matrix should be returned
instead of the RxODE's solved object}

\item{sigma}{Matrix for simulating residual variability for the number of observations.
     This uses the \code{\link[mvnfast]{rmvn}} or \code{\link[mvnfast]{rmvt}} assuming mean 0 and covariance given by this named
     matrix.  The residual-type variability is created for each of the named \"err\" components specified by the sigma matrix's
     column names. This then creates a random deviate that is used in the place of each named variable.  This should
     not really be used in the ODE specification.  If it is, though, it is treated as a time-varying covariate.

     Also note this does not use R's random numbers, rather it uses a cryptographic parallel random number generator;

     To allow reproducible research you must both set a random seed with R's \code{\link[base]{set.seed}} function, and keep the
     number of cores constant.  By changing either one of these variables, you will arrive at different random numbers.}

\item{amountUnits}{Dosing amount units.}

\item{timeUnits}{Time units.}

\item{covNames}{Covariate names in dataset.}

\item{df}{Degrees of freedom for \code{\link[mvnfast]{rmvt}}.  If \code{NULL}, or \code{Inf}, then use a normal distribution.  The
default is normal.}

\item{ncoresRV}{The number of cores for residual simulation.  By default, this is \code{1}.}

\item{isChol}{is a boolean indicating that the Cholesky decomposition of the \code{sigma} covariance matrix is supplied instead of
the  \code{sigma} matrix itself.}
}
\value{
A data structure to allow C-based for loop (ie solving each
      individual in C)
}
\description{
Setup a data frame for solving multiple subjects at once in RxODE.
}
\author{
Matthew L. Fidler
}
