% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{rxDataParSetup}
\alias{rxDataParSetup}
\title{Setup Data and Parameters}
\usage{
rxDataParSetup(object, params = NULL, events = NULL, inits = NULL,
  sigma = NULL, sigma.df = Inf, sigma.ncores = 1L, sigma.isChol = FALSE)
}
\arguments{
\item{object}{is a either a RxODE family of objects, or a file-name
with a RxODE model specification, or a string with a RxODE
model specification.}

\item{params}{a numeric named vector with values for every
parameter in the ODE system; the names must correspond to the
parameter identifiers used in the ODE specification;}

\item{events}{an \code{eventTable} object describing the input
(e.g., doses) to the dynamic system and observation sampling
time points (see \code{\link{eventTable}});}

\item{inits}{a vector of initial values of the state variables
(e.g., amounts in each compartment), and the order in this
vector must be the same as the state variables (e.g., PK/PD
compartments);}

\item{sigma}{Named sigma matrix.}
}
\description{
Setup Data and Parameters
}
\author{
Matthew L. Fidler
}
