% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxDelete}
\alias{rxDelete}
\title{Delete the dll for the model}
\usage{
rxDelete(obj)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
a boolean stating if the operation was successful.
}
\description{
This function deletes the dll, but doesn't delete the model
information in the object.
}

