% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{rxDemoteAddErr}
\alias{rxDemoteAddErr}
\title{Demote the error type}
\usage{
rxDemoteAddErr(errType)
}
\arguments{
\item{errType}{Error type factor}
}
\value{
Demoted Error Type
}
\description{
Demote the error type
}
\examples{
rxErrTypeCombine("add") \%>\%
  rxErrTypeCombine("prop")

# This removes the internal additive error
rxErrTypeCombine("add") \%>\%
  rxErrTypeCombine("prop") \%>\%
  rxDemoteAddErr()

# This is used for logitNorm(NA), the additive portion is stripped
}
\author{
Matthew Fidler
}
\keyword{internal}
