% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDetaDomega}
\alias{rxDetaDomega}
\title{Calculate d(eta)/d(omega)}
\usage{
rxDetaDomega(rho)
}
\arguments{
\item{rho}{environment where omegaInv.dOmega.omegaInv and tr.omegaInv.dOmega.0.5
are calculated.  This is done with the rxSymInv function.}

\item{eta}{the eta to caluclate the differential for.}
}
\value{
Nothing.  Add omega.28 and omega.47 to the environment rho.
}
\description{
Calculate d(eta)/d(omega)
}
\keyword{internal}
