% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxDll}
\alias{rxDll}
\alias{rxDll.RxODE}
\alias{rxDll.character}
\alias{rxDll.rxDll}
\title{Return the dll associated with the RxODE object}
\usage{
rxDll(obj, ...)

\method{rxDll}{character}(obj, ...)

\method{rxDll}{rxDll}(obj, ...)

\method{rxDll}{RxODE}(obj, ...)
}
\arguments{
\item{obj}{A RxODE family of objects or a character string of the
model specification or location of a file with a model
specification.}
}
\value{
a path of the library
}
\description{
This will return the dynamic load library or shared object used to
run the C code for RxODE.
}
\keyword{internal}

