% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxDllLoaded}
\alias{rxDllLoaded}
\title{Determine if the rxDll is loaded or not.}
\usage{
rxDllLoaded(x, retry = TRUE)
}
\arguments{
\item{x}{is a RxODE family of objects}
}
\value{
a boolean stating if the DLL is loaded
}
\description{
Determine if the rxDll is loaded or not.
}
\author{
Matthew L.Fidler
}
