% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RxODE.R
\name{rxDynLoad}
\alias{rxDynLoad}
\alias{rxLoad}
\title{Load RxODE object}
\usage{
rxDynLoad(obj)

rxLoad(obj)
}
\arguments{
\item{obj}{A RxODE family of objects}
}
\value{
Boolean returning if the RxODE library is loaded.
}
\description{
Load RxODE object
}
\author{
Matthew L.Fidler
}
\keyword{internal}
