% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxDynProtect}
\alias{rxDynProtect}
\title{Protect DLL(s) from unloading}
\usage{
rxDynProtect(dll)
}
\arguments{
\item{dlls}{is a list of DLLs that shouldn't be unloaded}
}
\description{
Protect DLL(s) from unloading
}
\author{
Matthew Fidler
}
