% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RxODE.R
\name{rxDynUnload}
\alias{rxDynUnload}
\alias{rxUnload}
\title{Unload RxODE object}
\usage{
rxDynUnload(obj)

rxUnload(obj)
}
\arguments{
\item{obj}{A RxODE family of objects}
}
\value{
Boolean returning if the RxODE library is loaded.
}
\description{
Unload RxODE object
}
\author{
Matthew L.Fidler
}
\keyword{internal}
