% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{rxEvid}
\alias{rxEvid}
\alias{as.rxEvid}
\alias{c.rxEvid}
\alias{[.rxEvid}
\alias{as.character.rxEvid}
\alias{format.rxEvid}
\alias{print.rxEvid}
\alias{[[.rxEvid}
\alias{units<-.rxEvid}
\alias{[<-.rxEvid}
\alias{type_sum.rxEvid}
\alias{pillar_shaft.rxEvid}
\alias{as.data.frame.rxEvid}
\alias{c.rxRateDur}
\alias{format.rxRateDur}
\title{EVID formatting for tibble and other places.}
\usage{
rxEvid(x)

as.rxEvid(x)

\method{c}{rxEvid}(x, ...)

\method{[}{rxEvid}(x, ...)

\method{as.character}{rxEvid}(x, ...)

\method{format}{rxEvid}(x, ...)

\method{print}{rxEvid}(x, ...)

\method{[[}{rxEvid}(x, ...)

\method{units}{rxEvid}(x) <- value

\method{[}{rxEvid}(x, i) <- value

type_sum.rxEvid(x)

pillar_shaft.rxEvid(x, ...)

\method{as.data.frame}{rxEvid}(x, row.names = NULL, optional = FALSE,
  ..., nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse =
  " "))

\method{c}{rxRateDur}(x, ...)

\method{format}{rxRateDur}(x, ...)
}
\arguments{
\item{x}{Item to be converted to a RxODE EVID specification.}

\item{...}{additional arguments to be passed to or from methods.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{nm}{Name of column in new data frame}
}
\description{
This is to make an EVID more readable by non
pharmacometricians. It displays what each means and allows it to
be displayed in a tibble.
}
\examples{

rxEvid(1:7)

}
