% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxExpandIfElse}
\alias{rxExpandIfElse}
\title{Expand if/else clauses into mutiple different types of lines.}
\usage{
rxExpandIfElse(model, removeInis = TRUE)
}
\arguments{
\item{model}{Model can be a character, or a RxODE model.  It needs
to have normalized syntax, that is \code{if (...)\{} has to be
on the same line.  The \code{else} statement must be on its
own line with the closing bracket of the \code{if} statement
on the previous line.  This \code{else} statment must also
contain the opening bracket, like the code \code{else \{}}

\item{removeInis}{When the model is an RxODE model, remove the
initialziation parameters from the model when \code{TRUE}, or
leave then when \code{FALSE}.}
}
\value{
A named character vector. The names of the vector are the
    logical conditions, the values are the lines that satisfy the
    logical conditions.
}
\description{
Expand if/else clauses into mutiple different types of lines.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
