% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiFinalizeLlik}
\alias{rxFoceiFinalizeLlik}
\title{Finalize Likelihood for individual.}
\usage{
rxFoceiFinalizeLlik(env)
}
\arguments{
\item{env}{Environment where likelihood is finalized.}
}
\value{
The likelihood with the fitted values, posthoc eta, and possibly the individual contribution to the gradient.
}
\description{
Finalize Likelihood for individual.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
