% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiInner}
\alias{rxFoceiInner}
\title{Solve the FOCEI inner problem}
\usage{
rxFoceiInner(object, ..., dv, eta, eta.bak = NULL, estimate = TRUE)
}
\arguments{
\item{object}{RxODE object}

\item{...}{values sent to rxFoceiEtaSetup and lbfgs}

\item{dv}{dependent variable}

\item{eta}{eta value}

\item{eta.bak}{backup eta value in case eta estimation fails}

\item{estimate}{Boolean indicating if the optimization should be
perfomed(TRUE), or just keep the eta, and calculate the Loglik
with fitted/posthoc attributes(FALSE).}
}
\value{
Loglik with fitted and posthoc attributes
}
\description{
Solve the FOCEI inner problem
}
\author{
Matthew L. Fidler
}
\keyword{internal}
