% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiLp}
\alias{rxFoceiLp}
\alias{rxFoceiLp.RxODE}
\alias{rxFoceiLp.rxDll}
\title{Get -Grad for individual}
\usage{
rxFoceiLp(object, ..., dv, eta, omegaInv)

\method{rxFoceiLp}{RxODE}(object, ..., dv, eta, omegaInv)

\method{rxFoceiLp}{rxDll}(object, ..., dv, eta, omegaInv)
}
\arguments{
\item{object}{RxODE object}

\item{...}{values sent to rxFoceiEtaSetup}

\item{dv}{dependent variable}

\item{eta}{eta values.}

\item{omegaInv}{Inverse omega}
}
\value{
-grad
}
\description{
For testing
}
\author{
Matthew L. Fidler
}
\keyword{internal}
