% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiTheta}
\alias{rxFoceiTheta}
\alias{rxFoceiTheta.rxFocei}
\alias{rxFoceiTheta.RxODE}
\alias{rxFoceiTheta.rxDll}
\title{FOCEI THETA setup,}
\usage{
rxFoceiTheta(object, ..., dv, eta, omegaInv, env, nonmem = FALSE)

\method{rxFoceiTheta}{rxFocei}(object, ..., dv, eta, omegaInv, env,
  nonmem = FALSE)

\method{rxFoceiTheta}{RxODE}(object, ..., dv, eta, omegaInv, env, nonmem)

\method{rxFoceiTheta}{rxDll}(object, ..., dv, eta, omegaInv, env, nonmem)
}
\arguments{
\item{object}{RxODE object}

\item{...}{values sent to rxFoceiEtaSetup}

\item{dv}{dependent variable}

\item{eta}{eta values.}

\item{omegaInv}{Inverse omega}

\item{env}{Environment to use, instead of rxFoceiEtaSetup environment}

\item{nonmem}{Match NONMEMs approximation of the hessian.}
}
\value{
environment of solved information.
}
\description{
This is basically for testing
}
\author{
Matthew L. Fidler
}
\keyword{internal}
