% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxFoExpandEta}
\alias{rxFoExpandEta}
\title{First Order Expansion of ETA}
\usage{
rxFoExpandEta(expr)
}
\arguments{
\item{expr}{RxODE model}
}
\value{
Return a RxODE model with first order Taylor expansion around ETA
}
\description{
First Order Expansion of ETA
}
\author{
Matthew L. Fidler
}
