% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxFromSymPy}
\alias{rxFromSymPy}
\title{Convert SymPy syntax to RxODE syntax}
\usage{
rxFromSymPy(text)
}
\arguments{
\item{text}{SymPy text to convert to RxODE syntax}
}
\value{
RxODE syntax
}
\description{
Convert SymPy syntax to RxODE syntax
}
\author{
Matthew L. Fidler
}
\keyword{internal}
