% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{rxGenSaem}
\alias{rxGenSaem}
\title{Generate pred-only SAEM RxODE model}
\usage{
rxGenSaem(obj, predfn, pkpars = NULL, sum.prod = FALSE, optExpression = TRUE)
}
\arguments{
\item{obj}{RxODE model (text or actual model)}

\item{predfn}{prediction function}

\item{pkpars}{PKpars function}
}
\value{
RxODE text
}
\description{
Generate pred-only SAEM RxODE model
}
\author{
Matt Fidler
}
