% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-focei.R
\name{rxGetDistributionFoceiLines}
\alias{rxGetDistributionFoceiLines}
\alias{rxGetDistributionFoceiLines.norm}
\alias{rxGetDistributionFoceiLines.t}
\alias{rxGetDistributionFoceiLines.default}
\alias{rxGetDistributionFoceiLines.rxUi}
\title{This is a S3 method for getting the distribution lines for a RxODE simulation}
\usage{
rxGetDistributionFoceiLines(line)

\method{rxGetDistributionFoceiLines}{norm}(line)

\method{rxGetDistributionFoceiLines}{t}(line)

\method{rxGetDistributionFoceiLines}{default}(line)

\method{rxGetDistributionFoceiLines}{rxUi}(line)
}
\arguments{
\item{line}{Parsed RxODE model environment}
}
\value{
Lines for the simulation of \code{ipred} and \code{dv}. This is based
on the idea that the focei parameters are defined
}
\description{
This is a S3 method for getting the distribution lines for a RxODE simulation
}
\author{
Matthew Fidler
}
\keyword{internal}
