% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err-sim.R
\name{rxGetDistributionSimulationLines}
\alias{rxGetDistributionSimulationLines}
\alias{rxGetDistributionSimulationLines.norm}
\alias{rxGetDistributionSimulationLines.t}
\alias{rxGetDistributionSimulationLines.default}
\alias{rxGetDistributionSimulationLines.rxUi}
\title{This is a S3 method for getting the distribution lines for a RxODE simulation}
\usage{
rxGetDistributionSimulationLines(line)

\method{rxGetDistributionSimulationLines}{norm}(line)

\method{rxGetDistributionSimulationLines}{t}(line)

\method{rxGetDistributionSimulationLines}{default}(line)

\method{rxGetDistributionSimulationLines}{rxUi}(line)
}
\arguments{
\item{line}{Parsed RxODE model environment}
}
\value{
Lines for the simulation of \code{ipred} and \code{dv}. This is based
on the idea that the focei parameters are defined
}
\description{
This is a S3 method for getting the distribution lines for a RxODE simulation
}
\author{
Matthew Fidler
}
\keyword{internal}
