% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxGetModel}
\alias{rxGetModel}
\title{Get model properties without compiling it.}
\usage{
rxGetModel(model, calcSens = FALSE, calcJac = FALSE)
}
\arguments{
\item{model}{RxODE specification}

\item{calcSens}{Calculate sensitivity vector/boolean}

\item{calcJac}{Calculate jacobian}
}
\value{
RxODE trans list
}
\description{
Get model properties without compiling it.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
