% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFormat.R
\name{rxHtml}
\alias{rxHtml}
\alias{rxHtml.rxSolve}
\title{Format rxSolve and related objects as html.}
\usage{
rxHtml(x, ...)

\method{rxHtml}{rxSolve}(x, ...)
}
\arguments{
\item{x}{RxODE object}

\item{...}{Extra arguments sent to kable}
}
\description{
Format rxSolve and related objects as html.
}
\author{
Matthew L. Fidler
}
