% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxIncludeDir}
\alias{rxIncludeDir}
\title{RxODE C headers include directory}
\usage{
rxIncludeDir(...)
}
\arguments{
\item{...}{Additional parameters sent to file.path}
}
\value{
RxODE include directory
}
\description{
Return the include directory
}
\details{
The include directory has the headers that may be needed to build
functions against the RxODE library.
}
\author{
Matthew L. Fidler
}
