% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxInits}
\alias{rxInit}
\alias{rxInits}
\title{Initial Values and State values for a RxODE object}
\usage{
rxInits(rxDllObj, vec, req, default = 0, noerror = FALSE, ...)

rxInit(rxDllObj, vec, req, default = 0, noerror = FALSE, ...)
}
\arguments{
\item{rxDllObj}{rxDll, RxODE, or named vector representing default
initial arguments}

\item{vec}{If supplied, named vector for the model.}

\item{req}{Required names, and the required order for the ODE solver}

\item{default}{a number or NA representing the default value for
parameters missing in \code{vec}, but required in \code{req}.}

\item{noerror}{is a boolean specifying if an error should be thrown
for missing parameter values when \code{default} = \code{NA}}
}
\description{
Returns the initial values of the rxDll object
}
\keyword{internal}

