% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxInvWishartVar}
\alias{rxInvWishartVar}
\title{Calculate Wishart Variance based on Omega matrix}
\usage{
rxInvWishartVar(Omega, nu)
}
\arguments{
\item{Omega}{is the square positive definite matrix}

\item{nu}{is the degrees of freedom of the Wishart Distribution.}
}
\value{
Variance matrix
}
\description{
Calculate Wishart Variance based on Omega matrix
}
