% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxIs}
\alias{rxIs}
\title{Check the type of an object using Rcpp}
\usage{
rxIs(obj, cls)
}
\arguments{
\item{obj}{Object to check}

\item{cls}{Type of class.  Only s3 classes and primitive classes are checked.
For matrix types they are distinguished as \code{numeric.matrix}, \code{integer.matrix},
\code{logical.matrix}, and \code{character.matrix} as well as the traditional \code{matrix}
class.}
}
\value{
A boolean indicating if the object is a member of the class.
}
\description{
Check the type of an object using Rcpp
}
\keyword{internal}
