% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxIsCurrent}
\alias{rxIsCurrent}
\title{Checks if the RxODE object was built with the current build}
\usage{
rxIsCurrent(obj)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
boolean indicating if this was built with current RxODE
}
\description{
Checks if the RxODE object was built with the current build
}
