% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RxODE.R
\name{rxIsLoaded}
\alias{rxIsLoaded}
\alias{rxDllLoaded}
\title{Determine if the DLL associated with the RxODE object is loaded}
\usage{
rxIsLoaded(obj)

rxDllLoaded(obj)
}
\arguments{
\item{obj}{A RxODE family of objects}
}
\value{
Boolean returning if the RxODE library is loaded.
}
\description{
Determine if the DLL associated with the RxODE object is loaded
}
\author{
Matthew L.Fidler
}
\keyword{internal}
