% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{rxIsNotebook}
\alias{rxIsNotebook}
\title{Return if the current document is a Rstudio notebook}
\usage{
rxIsNotebook()
}
\value{
If we are in a notebook
}
\description{
Return if the current document is a Rstudio notebook
}
\examples{

rxIsNotebook() # Check to see if you are running in a notebook.

}
\author{
Max Gordon
}
\keyword{internal}
