% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxKahanSum}
\alias{rxKahanSum}
\title{Using the Kahan method, take a more accurate sum}
\usage{
rxKahanSum(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Sum of numbers
}
\description{
Using the Kahan method, take a more accurate sum
}
\references{
\url{https://en.wikipedia.org/wiki/Kahan_summation_algorithm}
}
