% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxLinCmt.R
\name{rxLinCmt}
\alias{rxLinCmt}
\title{Interface to solved linear compartments implemented in RxODE}
\usage{
rxLinCmt(t = NULL, params = NULL, events = NULL, cmt = 1)
}
\arguments{
\item{t}{Times to Solve}

\item{params}{Parameters for solved linear compartments}

\item{events}{Event Table for solving}

\item{cmt}{The compartment for the solving, by default this is one}
}
\value{
A matrix with the solved equations and the derivatives of
    the equation based on the parameters you provide.
}
\description{
Interface to solved linear compartments implemented in RxODE
}
\author{
Matthew L. Fidler
}
