% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxLinCmt.R
\name{rxLinCmtTrans}
\alias{rxLinCmtTrans}
\title{This translates the parameters specified by the model in the correct type of solving.}
\usage{
rxLinCmtTrans(modText)
}
\arguments{
\item{modText}{model text}
}
\value{
Translated solve for RxODE linear compartments
}
\description{
This translates the parameters specified by the model in the correct type of solving.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
