% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxLoad}
\alias{rxLoad}
\title{Load the dll for the object}
\usage{
rxLoad(obj)
}
\arguments{
\item{obj}{a RxODE family of objects}
}
\description{
This loads the dll into the current R session to allow C functions
to be called in R.
}

