% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxLoadDir}
\alias{rxLoadDir}
\title{RxODE load directory}
\usage{
rxLoadDir(...)
}
\arguments{
\item{...}{File name under load directory, or nothing for load
directory alone.}
}
\value{
RxODE load Directory
}
\description{
RxODE load directory
}
\author{
Matthew L. Fidler
}
\keyword{internal}
