% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxLock}
\alias{rxLock}
\alias{rxUnlock}
\title{Lock/unlocking of RxODE dll file}
\usage{
rxLock(obj)

rxUnlock(obj)
}
\arguments{
\item{obj}{A RxODE family of objects}
}
\description{
Lock/unlocking of RxODE dll file
}
