% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxLockFree}
\alias{rxLockFree}
\title{Lock the freeing routines}
\usage{
rxLockFree(lock)
}
\arguments{
\item{lock}{This determines if the freeing is stopped}
}
\description{
This makes sure that the freeing routines are not called.
}
\keyword{internal}
