% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxLogicToVar.R
\name{rxLogicToVar}
\alias{rxLogicToVar}
\title{Change binary logical expressions to variables}
\usage{
rxLogicToVar(x)
}
\arguments{
\item{x}{RxODE model that can be accessed by rxNorm}
}
\value{
list of logical expressions and replaced variables
}
\description{
This is meant to be used with non-branching code.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
