% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxLogifyModel}
\alias{rxLogifyModel}
\title{Takes a model and expands it to log multiplication}
\usage{
rxLogifyModel(model, expand = TRUE)
}
\arguments{
\item{model}{RxODE model}

\item{expand}{Expand the symbolic expression using SymPy?}
}
\value{
Lines for expanded model. (but not compiled)
}
\description{
This is a numerical trick to help reduce multiplcation errors when
numbers are of very different magnitude.
}
\author{
Matthew L. Fidler
}
