% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{rxMatrix}
\alias{rxMatrix}
\title{Specify lower triangular matrix}
\usage{
rxMatrix(x, ...)
}
\arguments{
\item{x}{list, matrix or expression, see details}

\item{...}{Other arguments treated as a list that will be
concatenated then reapplied to this function.}
}
\value{
Matrix
}
\description{
Specify lower triangular matrix
}
