% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxMd5}
\alias{rxMd5}
\title{Return the md5 of an RxODE object or file}
\usage{
rxMd5(model, extraC = NULL, ...)
}
\arguments{
\item{model}{either a string representing the path to a model file
or a RxODE object}

\item{extraC}{The extra C file used to help create the md5.  If the
model is an RxODE object, this argument is ignored.}

\item{...}{ignored arguments}
}
\value{
If this is a RxODE object, return a named list:

\item{\code{file_md5}}{ is the model's file's md5}

\item{\code{parsed_md5}}{ is the parsed model's file's md5.}

Otherwise return the md5 based on the arguments provided
}
\description{
This md5 is based on the model and possibly the extra c code
supplied for the model
}
\keyword{internal}

