% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxModels_}
\alias{rxModels_}
\title{Get the rxModels environment where loaded RxODE models are stored.}
\usage{
rxModels_()
}
\description{
Get the rxModels environment where loaded RxODE models are stored.
}
\keyword{internal}
