% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.R
\name{rxMuRef}
\alias{rxMuRef}
\title{Get mu-referencing model from model variables}
\usage{
rxMuRef(mod, ini = NULL)
}
\arguments{
\item{mod}{Model}

\item{theta}{Thetas from model initialization (like Lori)}

\item{eta}{Etas from model initialization (like lotri)}
}
\value{

}
\description{
The rxMuRef is the core of the nlmixr ui functions
}
\details{
This function takes the initialization values from \code{lotri()} the
parsed RxODE model to generate mu-referenced models adding
mu-references for etas that do not have them to allow saem to
support non mu-referenced models by a parsing trick.
}
\examples{

# First lets get a lotri initialization block:

ini <- lotri({
   ## You may label each parameter with a comment
   tka <- 0.45 # Log Ka
   tcl <- log(c(0, 2.7, 100)) # Log Cl
   ## This works with interactive models
   ## You may also label the preceding line with label("label text")
   tv <- 3.45; label("log V")
   ## the label("Label name") works with all models
   eta.ka ~ 0.6
   eta.cl + eta.v ~ c(0.3,
                      0.001, 0.1)
   add.sd <- 0.7
})


ini2 <- lotri({
       ## You may label each parameter with a comment
       tka <- 0.45 # Log Ka
       tcl <- log(c(0, 2.7, 100)) # Log Cl
       ## This works with interactive models
       ## You may also label the preceding line with label("label text")
       tv <- 3.45; label("log V")
       add.sd <- 0.7
    })

}
\author{
Matthew Fidler
}
