% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxNeumaierSum}
\alias{rxNeumaierSum}
\title{Using the Neumaier method, take a more accurate sum}
\usage{
rxNeumaierSum(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Sum of numbers, a bit more accurate than rxKahanSum
}
\description{
Using the Neumaier method, take a more accurate sum
}
\references{
\url{https://en.wikipedia.org/wiki/Kahan_summation_algorithm}
}
