% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxNorm}
\alias{rxNorm}
\title{Get the normalized model}
\usage{
rxNorm(obj)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
Normalized Normal syntax (no comments)
}
\description{
This get the syntax prefered model for processing
}
\author{
Matthew L. Fidler
}
