% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxodeGc}
\alias{rxodeGc}
\title{Garbage Collection for RxODE objects}
\usage{
rxodeGc(env)
}
\arguments{
\item{env}{RxODE environment for cleanup.}
}
\value{
Nothing
}
\description{
When objects are created without a model name/DLL name and RxODE
option RxODE.delete.unnamed is TRUE, remove the associated DLL
when R is done with them (i.e. when the RxODE object is deleted or
R exits normally.)  If R crashes, thes files will still remain.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
