% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxValidate.R
\name{rxodeTest}
\alias{rxodeTest}
\title{Wrap a test}
\usage{
rxodeTest(code, test = "cran")
}
\arguments{
\item{type}{Type of test or filter of test type}

\item{check}{Use devtools::check instead}
}
\value{

}
\description{
.. content for \details{} ..
}
\author{
Matthew Fidler
}
