% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxPairwiseSum}
\alias{rxPairwiseSum}
\title{Return an accurate floating point sum of values}
\usage{
rxPairwiseSum(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
A sum of numbers with a rounding error of O(lg n) instead
    of O(n).
}
\description{
This was taken by NumPy and adapted for use here.
}
\references{
\url{https://github.com/juliantaylor/numpy/blob/b0bc01275cac04483e6df021211c1fa2ba65eaa3/numpy/core/src/umath/loops.c.src}

\url{https://github.com/numpy/numpy/pull/3685}
}
\author{
Matthew Fidler (R implementation), Julian Taylor, Nathaniel
    J Smith, and others in NumPy team.
}
