% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxParams}
\alias{rxParams}
\title{Parameters specified by the model}
\usage{
rxParams(obj, ...)
}
\arguments{
\item{obj}{is a RxODE family of objects}
}
\value{
a character vector listing the parameters in the model.
}
\description{
This return the model's parameters that are required to solve the
ODE system.
}

