% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R, R/symengine.R
\name{rxParsePred}
\alias{rxParsePred}
\title{Prepare Pred function for inclusion in RxODE}
\usage{
rxParsePred(x, init = NULL, err = NULL)

rxParsePred(x, init = NULL, err = NULL)
}
\arguments{
\item{x}{pred function}

\item{init}{Initilization vector}

\item{x}{pred function}
}
\value{
RxODE transformed text.

RxODE transformed text.
}
\description{
Prepare Pred function for inclusion in RxODE

Prepare Pred function for inclusion in RxODE
}
\author{
Matthew L. Fidler

Matthew L. Fidler
}
\keyword{internal}
