% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE-win-setup.R
\name{rxPhysicalDrives}
\alias{rxPhysicalDrives}
\title{Returns a list of physical drives that have been or currently are
mounted to the computer.}
\usage{
rxPhysicalDrives(duplicates = FALSE)
}
\arguments{
\item{duplicates}{Return drives with duplicate entries in
\code{SYSTEM\\MountedDevices}; These are likely removable media.  By default this is \code{FALSE}}
}
\value{
Drives with letters
}
\description{
This excludes network drives.  See
\url{https://www.forensicmag.com/article/2012/06/windows-7-registry-forensics-part-5}
}
\author{
Matthew L. Fidler
}
\keyword{internal}
