% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\name{rxPkg}
\alias{rxPkg}
\title{Creates a package from compiled RxODE models}
\usage{
rxPkg(
  ...,
  package,
  wd = getwd(),
  action = c("install", "build", "binary", "create"),
  fields = list()
)
}
\arguments{
\item{...}{Models to build a package from}

\item{package}{String of the package name to create}

\item{action}{Type of action to take after package is created}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}
}
\description{
Creates a package from compiled RxODE models
}
\author{
Matthew Fidler
}
