% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxPlot}
\alias{rxPlot}
\title{Plot a model digram for simple ODE models}
\usage{
rxPlot(RxODEobj, family = "sans", interactive = FALSE, shape = c("square",
  "circle", "csquare", "rectangle", "crectangle", "vrectangle", "sphere",
  "none"), size = 30, colors = c(accent1 = "#0460A9"),
  fillColor = "accent1", font = 2, labelColor = "white",
  lineColor = "accent1", shapeEnd = c("none", "sphere", "circle", "square",
  "csquare", "rectangle", "crectangle", "vrectangle"), sizeEnd = 10, ...)
}
\arguments{
\item{RxODEobj}{RxODE object}

\item{family}{Font family to use}

\item{interactive}{If true, use Tcl/tk to plot the diagram and
allow movement of the compartments}

\item{shape}{Box shape, by default it is a square}

\item{size}{size of square for compartments}

\item{colors}{is a named list of colors.  If any of the names match
the colors specified by this function, it will use the color in
this list.  For example \code{accent1} would match the default
color in this list.}

\item{fillColor}{Color of the fill}

\item{font}{Font (1: plain, 2: bold, 3: italic, 4: bold/italic,
5:symbol)}

\item{labelColor}{Color of the label}

\item{lineColor}{Color of the line}

\item{shapeEnd}{End shape}

\item{sizeEnd}{Size of end}

\item{...}{Other arguments passed to igraph's plot method.

Currently this method supports simple elimination type models and
indirect-effect models.  Michelis Menton models are not yet
supported.}
}
\description{
This plots a line diagram for the current system of differential
equations.
}

