% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{rxPreferredDistributionName}
\alias{rxPreferredDistributionName}
\title{Change distribution name to the preferred distribution name term}
\usage{
rxPreferredDistributionName(dist)
}
\arguments{
\item{dist}{This is the input distribution}
}
\value{
Preferred distribution term
}
\description{
This is determined by the internal preferred condition name list
\code{.errIdenticalDists}
}
\examples{

rxPreferredDistributionName("dnorm")

rxPreferredDistributionName("add")

}
\author{
Matthew Fidler
}
