% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxProd}
\alias{rxProd}
\title{Using RxODE's default method, take a product}
\usage{
rxProd(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Product of numbers
}
\description{
Using RxODE's default method, take a product
}
