% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxrandom.R
\name{rxRandNV}
\alias{rxRandNV}
\title{Create a random "normal" matrix using vandercorput generator}
\usage{
rxRandNV(nrow = 1, ncol = 1)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of Columns}
}
\value{
Matrix of random numbers
}
\description{
Create a random "normal" matrix using vandercorput generator
}
\examples{

rxRandNV(1,1)
rxRandNV(3,2)

}
\author{
Matthew Fidler
}
