% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxReq}
\alias{rxReq}
\title{Require namespace, otherwise throw error.}
\usage{
rxReq(pkg)
}
\arguments{
\item{pkg}{Package required for function to work.}
}
\value{
Nothing
}
\description{
Require namespace, otherwise throw error.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
