% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxReservedKeywords.R
\docType{data}
\name{rxReservedKeywords}
\alias{rxReservedKeywords}
\title{A list and description of Rode supported syntax functions}
\format{
A data frame with 3 columns and 98 or more rows
\describe{
\item{Function}{Function Name}
\item{Description}{Function Description}
\item{Aliases}{Function Alaises}
}
}
\usage{
rxReservedKeywords
}
\description{
A list and description of Rode supported syntax functions
}
\keyword{datasets}
