% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxRmIni}
\alias{rxRmIni}
\title{Remove INIs}
\usage{
rxRmIni(x)
}
\arguments{
\item{x}{RxODE list of lines to remove}
}
\value{
RxODE lines with inis removed.
}
\description{
Remove INIs
}
\author{
Matthew L. Fidler
}
\keyword{internal}
