% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxS}
\alias{rxS}
\title{Load a model into a symengine environment}
\usage{
rxS(x, doConst = TRUE)
}
\arguments{
\item{x}{RxODE object}
}
\value{
RxODE/symengine environment
}
\description{
Load a model into a symengine environment
}
\author{
Matthew Fidler
}
