% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxSetIni0}
\alias{rxSetIni0}
\title{Set Initial conditions to time zero instead of the first observed/dosed time}
\usage{
rxSetIni0(ini0 = TRUE)
}
\arguments{
\item{ini0}{When TRUE (default), set initial conditions to time
zero. Otherwise the initial conditions are the first observed
time.}
}
\description{
Set Initial conditions to time zero instead of the first observed/dosed time
}
