% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxSetProd}
\alias{rxSetProd}
\title{Defunct setting of product}
\usage{
rxSetProd(type = c("long double", "double", "logify"))
}
\arguments{
\item{type}{used to be type of product}
}
\value{
nothing
}
\description{
Defunct setting of product
}
