% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxSetProgressBar}
\alias{rxSetProgressBar}
\title{Turn on or off the progress bar}
\usage{
rxSetProgressBar(on = TRUE)
}
\arguments{
\item{on}{When TRUE (default), the progress bar is turned on.}
}
\description{
Turn on or off the progress bar
}
