% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxSetSilentErr}
\alias{rxSetSilentErr}
\title{Silence some of RxODE's C/C++ messages}
\usage{
rxSetSilentErr(silent)
}
\arguments{
\item{silent}{can be 0L "noisy"  or 1L "silent"}
}
\value{
TRUE; called for side effects
}
\description{
Silence some of RxODE's C/C++ messages
}
\keyword{internal}
