% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxSetSum}
\alias{rxSetSum}
\title{Defunct setting of sum}
\usage{
rxSetSum(type = c("pairwise", "fsum", "kahan", "neumaier", "c"))
}
\arguments{
\item{type}{used to be type of product}
}
\value{
nothing
}
\description{
Defunct setting of sum
}
