% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxSetupIni}
\alias{rxSetupIni}
\title{Setup the initial conditions.}
\usage{
rxSetupIni(obj, inits = NULL)
}
\arguments{
\item{obj}{RxODE object}

\item{inits}{A numeric vector of initial conditions.}
}
\value{
initial conditions that were setup
}
\description{
Setup the initial conditions.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
