% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode-options.R
\name{rxSetupMemoize}
\alias{rxSetupMemoize}
\title{This setups the memoized functions.}
\usage{
rxSetupMemoize()
}
\description{
To easily create a memozied function by adding a \code{.slow <- NULL}
to the end of a function.
}
\details{
For example, to memozie the function in the namespace
\code{rxModelVars.character} you would add a line:
\code{rxModelVars.character.slow <- NULL}
}
\author{
Matthew L. Fidler
}
