% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxSetupScale}
\alias{rxSetupScale}
\title{Setup the initial conditions.}
\usage{
rxSetupScale(obj, scale = NULL, extraArgs = NULL)
}
\arguments{
\item{obj}{RxODE object}

\item{extraArgs}{A list of extra args to parse for initial conditions.}

\item{inits}{A numeric vector of initial conditions.}
}
\description{
Setup the initial conditions.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
