% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxShiny.R
\name{rxShiny}
\alias{rxShiny}
\alias{rxShiny.default}
\title{Use Shiny to help develop an RxODE model}
\usage{
rxShiny(object, ...)


  \method{rxShiny}{default}(object = "\\nC2 = centr/V2\\nC3 = peri/V3\\nd/dt(depot) =-KA*depot\\nd/dt(centr) = KA*depot - CL*C2 - Q*C2 + Q*C3\\nd/dt(peri)  =                    Q*C2 - Q*C3\\nd/dt(eff)  = Kin - Kout*(1-C2/(EC50+C2))*eff\\n",
  ...)
}
\arguments{
\item{object}{A RxODE family of objects. If not supplied a
2-compartment indirect effect model is used.  If it is
supplied, use the model associated with the RxODE object for
the model exploration.}

\item{...}{Other arguments passed to rxShiny.  Currently doesn't
do anything.}
}
\value{
Nothing; Starts a shiny server
}
\description{
Use Shiny to help develop an RxODE model
}
\author{
Zufar Mulyukov and Matthew L. Fidler
}
