% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxSigmoidInfo}
\alias{rxSigmoidInfo}
\title{Internal function to detect sigmoid functions}
\usage{
rxSigmoidInfo(text = "+Kin-Kout*(1-centr/V2/(EC50+centr/V2))*eff")
}
\description{
Used in plotting to figure out if there is a sigmoid function
}
\author{
Matthew L.Fidler
}
\keyword{internal}

