% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxSolveFree}
\alias{rxSolveFree}
\title{Free the C solving/parsing information.}
\usage{
rxSolveFree()
}
\value{
logical indicating if the memory was successfully freed
}
\description{
Take the ODE C system and free it.
}
\keyword{internal}
