% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxSplitPlusQ}
\alias{rxSplitPlusQ}
\title{This function splits a function based on + or - terms}
\usage{
rxSplitPlusQ(x, level = 0, mult = FALSE)
}
\arguments{
\item{x}{Quoted R expression for splitting}

\item{level}{Internal level of parsing}

\item{mult}{boolean to split based on * and / expressions instead.
By default this is turned off.}
}
\value{
character vector of the split expressions
}
\description{
It uses the parser and does not disturb terms within other
functions.  For example:
}
\details{
a\emph{exp(b+c)+d}log(e-f)-g*f

would return

c("a * exp(b + c)", "d * log(e - f)", "- g * f")
}
\author{
Matthew L. Fidler
}
\keyword{internal}
