% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxStack}
\alias{rxStack}
\title{Stack a solved object for things like ggplot}
\usage{
rxStack(Data, vars = NULL)
}
\arguments{
\item{Data}{is a RxODE object to be stacked.}

\item{vars}{Variables to include in stacked data; By default this
is all the variables when vars is NULL.}
}
\value{
Stacked data with value and trt, where value is the values
  and trt is the state and lhs variables.
}
\description{
Stack a solved object for things like ggplot
}
\author{
Matthew Fidler
}
