% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxState}
\alias{rxState}
\title{State variables}
\usage{
rxState(obj, state, ...)
}
\arguments{
\item{obj}{is a RxODE family of objects}

\item{state}{is a string indicating the state or compartment that
you would like to lookup.}

\item{...}{Ignored arguments}
}
\value{
If state is missing, return a character vector of all the states.

If state is a string, return the compartment number of the named state.
}
\description{
This returns the model's compartments or states.
}
\seealso{
\code{\link{RxODE}}
}

