% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxStrict}
\alias{rxStrict}
\title{Strict RxODE sytax options}
\usage{
rxStrict(..., silent = FALSE)
}
\arguments{
\item{...}{other arguments.  If the first argument is an
expression, then evaluate this expression under rxStrict and
then restore the user's options.  If the first argument is
logical, assume it is the silent argument.}

\item{silent}{when true, also silence the syntax errors and
interactive output (useful in testing).}
}
\description{
This sets the RxODE syntax to be strict for backward compatability.
}
\author{
Matthew L. Fidler
}

