% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxSum}
\alias{rxSum}
\title{Using RxODE's default method, take a sum}
\usage{
rxSum(numbers)
}
\arguments{
\item{numbers}{A vector of numbers to sum.}
}
\value{
Sum of numbers
}
\description{
Using RxODE's default method, take a sum
}
