% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{rxSumProdModel}
\alias{rxSumProdModel}
\title{Recast model in terms of sum/prod}
\usage{
rxSumProdModel(model)
}
\arguments{
\item{model}{RxODE model}
}
\value{
model string with prod(.) and sum(.) for all these
    operations.
}
\description{
Recast model in terms of sum/prod
}
\author{
Matthew L. Fidler
}
