% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxSupportedFuns}
\alias{rxSupportedFuns}
\title{Get list of supported functions}
\usage{
rxSupportedFuns()
}
\value{
list of supported functions in RxODE
}
\description{
Get list of supported functions
}
\examples{
rxSupportedFuns()
}
