% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxSymInv.R
\name{rxSymDiag}
\alias{rxSymDiag}
\title{Get the theta numbers for the diagonal elements.}
\usage{
rxSymDiag(x)
}
\arguments{
\item{x}{object created with rxSymInvCreate}
}
\value{
Theta numbers for diagonal elements
}
\description{
These start at element 1 like in R instead of element 0 like in C
}
\author{
Matthew L. Fidler
}
