% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxSymInv.R
\name{rxSymInvCholCreate}
\alias{rxSymInvCholCreate}
\title{Creates an object for calculating Omega/Omega^-1 and derivatives}
\usage{
rxSymInvCholCreate(mat, diag.xform = c("log", "sqrt", "identity"),
  create.env = TRUE)
}
\arguments{
\item{mat}{Initial Omega matrix}

\item{diag.xform}{transformation to diagonal elements of OMEGA. or chol(Omega^-1)}

\item{create.env}{-- Create an environment to calculate the inverses.}
}
\value{
A rxSymInv object
}
\description{
Creates an object for calculating Omega/Omega^-1 and derivatives
}
\author{
Matthew L. Fidler
}
\keyword{internal}
