% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxSymInv.R
\name{rxSymInvCholN}
\alias{rxSymInvCholN}
\title{Return the dimension of the built-in derivatives/inverses}
\usage{
rxSymInvCholN()
}
\description{
Return the dimension of the built-in derivatives/inverses
}
\keyword{internal}
