% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPy}
\alias{rxSymPy}
\title{Execute sympy statement.}
\usage{
rxSymPy(...)
}
\arguments{
\item{...}{Parameters sent to Jython to execute.}
}
\value{
String representing the return from sympy environment
}
\description{
Execute sympy statement.
}
\author{
G Grothendieck, Matthew L. Fidler
}
\keyword{internal}
