% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyClear}
\alias{rxSymPyClear}
\title{Delete variable if exists.}
\usage{
rxSymPyClear(var)
}
\arguments{
\item{var}{Variable to delete.}
}
\description{
Delete variable if exists.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
