% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyDfDy}
\alias{rxSymPyDfDy}
\title{Calculate df/dy derivatives}
\usage{
rxSymPyDfDy(model, df, dy, vars = FALSE)
}
\arguments{
\item{model}{RxODE models}

\item{df}{is a string for the state in the df(.)/dy(.).  If
missing and dy is missing, all the df(.)/dy(.) components are
calculated according to the \code{vars} parameter below.}

\item{dy}{is a string for the state or variable in the df(.)/dy(.).}

\item{vars}{is a boolean indicating if parameters will be included
    for the dy component in the df(.)/dy(.), instead of just state
    variables (required for sensitivity equations).

    vars can also be a list of variables.  In this case the
    df(.)/dy(.) will be augmented with the variables in this list.}
}
\value{
RxODE syntax lines for the df(.)/dy(.)
}
\description{
Calculate df/dy derivatives
}
\author{
Matthew L. Fidler
}
\keyword{internal}
