% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyExec}
\alias{rxSymPyExec}
\title{Execute python statement without getting the return value.}
\usage{
rxSymPyExec(...)
}
\arguments{
\item{...}{Parameters sent to Jython to execute.}
}
\value{
Nothing
}
\description{
Execute python statement without getting the return value.
}
\author{
G Grothendieck, Matthew L. Fidler
}
\keyword{internal}
