% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyExists}
\alias{rxSymPyExists}
\title{Does the varaible exists in the sympy python environment}
\usage{
rxSymPyExists(var)
}
\arguments{
\item{var}{Variable to test if it exists.}
}
\value{
boolean
}
\description{
Does the varaible exists in the sympy python environment
}
\author{
Matthew L. Fidler
}
\keyword{internal}
