% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyExpand}
\alias{rxSymPyExpand}
\title{Return the expanded expression (via sympy)}
\usage{
rxSymPyExpand(x, expr = "expand")
}
\arguments{
\item{x}{text sympy-compatible expression}

\item{expr}{SymPy expression to use.  By default uses expand.}
}
\value{
sympy value for rxSymPy...
}
\description{
Return the expanded expression (via sympy)
}
\author{
Matthew L. Fidler
}
\keyword{internal}
