% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyFix}
\alias{rxSymPyFix}
\title{Fix SymPy expressions to be R parsable expressions}
\usage{
rxSymPyFix(var)
}
\arguments{
\item{var}{sympy expression}
}
\value{
R valid expression
}
\description{
Fix SymPy expressions to be R parsable expressions
}
\author{
Matthew L. Fidler
}
