% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyFunctions}
\alias{rxSymPyFunctions}
\title{Setup sympy functions}
\usage{
rxSymPyFunctions(functions)
}
\arguments{
\item{functions}{a list of functions}
}
\description{
This creates sympy unspecified functions for later evaulation in the CAS sympy.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
