% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyLincmtDvdx}
\alias{rxSymPyLincmtDvdx}
\title{This creates the dv/dx RxODE model for a linear solved system}
\usage{
rxSymPyLincmtDvdx(model, ncmt, parameterization, optExpression = TRUE)
}
\arguments{
\item{model}{RxODE pred model}

\item{ncmt}{Number of compartments of the solved system}

\item{parameterization}{Integer representing parameterization type}

\item{optExpression}{boolean indicating if you should optimize the
expression.}
}
\value{
RxODE model expressing \code{dvdx}
}
\description{
This creates the dv/dx RxODE model for a linear solved system
}
\author{
Matthew L. Fidler
}
\keyword{internal}
