% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyReserved}
\alias{rxSymPyReserved}
\title{Return a list of reserved functions and variables from sympy}
\usage{
rxSymPyReserved()
}
\value{
List of reserevd functions and variaibles from sympy.
}
\description{
Return a list of reserved functions and variables from sympy
}
\author{
Matthew L. Fidler
}
\keyword{internal}
