% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySensitivity}
\alias{rxSymPySensitivity}
\title{Calculate the sensitivity equations for a model}
\usage{
rxSymPySensitivity(model)
}
\arguments{
\item{model}{RxODE family of objects}
}
\value{
Model syntax that includes the sensitivity parameters.
}
\description{
This expands the model to calculate sensitivities.  This requires
rSymPy.
}
\author{
Matthew L. Fidler
}
