% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetup}
\alias{rxSymPySetup}
\title{Setup a sympy environment that sets up the specified RxODE model.}
\usage{
rxSymPySetup(model)
}
\arguments{
\item{model}{RxODE family of objects}
}
\description{
Setup a sympy environment that sets up the specified RxODE model.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
