% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetupIf}
\alias{rxSymPySetupIf}
\title{Setup SymPy environment if needed}
\usage{
rxSymPySetupIf(model)
}
\arguments{
\item{model}{RxODE lines to setup}
}
\value{
model lines
}
\description{
Setup SymPy environment if needed
}
\author{
Matthew L. Fidler
}
\keyword{internal}
