% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetupPred}
\alias{rxSymPySetupPred}
\title{Setup Pred function based on RxODE object.}
\usage{
rxSymPySetupPred(obj, predfn, pkpars, errfn)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}
}
\value{
RxODE object expanded with predfn and with calculated sensitivities.
}
\description{
This is for the so-called inner problem.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
