% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyVars}
\alias{rxSymPyVars}
\title{Setup sympy variables}
\usage{
rxSymPyVars(model)
}
\arguments{
\item{model}{RxODE family of objects}
}
\description{
This creates sympy variables for later evaulation in the CAS sympy.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
