% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyVersion}
\alias{rxSymPyVersion}
\title{Return the version of Sympy that is running}
\usage{
rxSymPyVersion()
}
\value{
Verson of sympy that is running.
}
\description{
Return the version of Sympy that is running
}
\author{
Matthew L. Fidler
}
