% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxSyntaxFunctions.R
\docType{data}
\name{rxSyntaxFunctions}
\alias{rxSyntaxFunctions}
\title{A list and description of Rode supported syntax functions}
\format{
A data frame with 3 columns and 98 or more rows
\describe{
\item{Reserved Name}{Reserved variable Name}
\item{Description}{Description of variable}
\item{Aliases}{Variable Aliases}
}
}
\usage{
rxSyntaxFunctions
}
\description{
A list and description of Rode supported syntax functions
}
\keyword{datasets}
