% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSyPyAddVars}
\alias{rxSyPyAddVars}
\title{Add undefined variables to SymPy}
\usage{
rxSyPyAddVars(txt)
}
\arguments{
\item{txt}{Text to add undefined variables to...}
}
\value{
nothing
}
\description{
Add undefined variables to SymPy
}
\author{
Matthew L. Fidler
}
\keyword{internal}
