% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rxTheme}
\alias{rxTheme}
\title{rxTheme is the RxODE theme for plots}
\usage{
rxTheme(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  grid = TRUE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{grid}{a Boolean indicating if the grid is on (\code{TRUE}) or off
(\code{FALSE}). This could also be a character indicating \code{x} or \code{y}.}
}
\value{
ggplot2 theme used in RxODE
}
\description{
rxTheme is the RxODE theme for plots
}
