% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxThetaEta}
\alias{rxThetaEta}
\title{Format theta and eta for parameter estimate values in RxODE}
\usage{
rxThetaEta(theta = NULL, eta = NULL)
}
\arguments{
\item{theta}{A vector of theta estimates}

\item{eta}{A vector of eta estimates}
}
\value{
A named vector for initial values.
}
\description{
Format theta and eta for parameter estimate values in RxODE
}
\author{
Matthew L. Fidler
}
\keyword{internal}
