% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxToSE}
\alias{rxToSE}
\alias{rxFromSE}
\alias{.rxFromSE}
\title{RxODE to symengine.R}
\usage{
rxToSE(x, envir = NULL)

rxFromSE(x)

.rxFromSE(x)
}
\arguments{
\item{x}{expression}

\item{envir}{\code{NULL}}
}
\description{
RxODE to symengine.R
}
\author{
Matthew L. Fidler
}
