% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxToSymPy}
\alias{rxToSymPy}
\title{Converts model specification into a SymPy model lines}
\usage{
rxToSymPy(model)
}
\arguments{
\item{model}{RxODE family of objects}
}
\value{
Lines for sympy
}
\description{
Converts model specification into a SymPy model lines
}
\author{
Matthew L. Fidler
}
\keyword{internal}
