% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxUiGet.R
\name{rxUiGet}
\alias{rxUiGet}
\alias{rxUiGet.theta}
\alias{rxUiGet.omega}
\alias{rxUiGet.muRefTable}
\alias{rxUiGet.multipleEndpoint}
\alias{rxUiGet.funPrint}
\alias{rxUiGet.fun}
\alias{rxUiGet.ini}
\alias{rxUiGet.iniFun}
\alias{rxUiGet.modelFun}
\alias{rxUiGet.default}
\title{S3 for getting information from UI model}
\usage{
rxUiGet(x, ...)

\method{rxUiGet}{theta}(x, ...)

\method{rxUiGet}{omega}(x, ...)

\method{rxUiGet}{muRefTable}(x, ...)

\method{rxUiGet}{multipleEndpoint}(x, ...)

\method{rxUiGet}{funPrint}(x, ...)

\method{rxUiGet}{fun}(x, ...)

\method{rxUiGet}{ini}(x, ...)

\method{rxUiGet}{iniFun}(x, ...)

\method{rxUiGet}{modelFun}(x, ...)

\method{rxUiGet}{default}(x, ...)
}
\arguments{
\item{x}{list of (UIenvironment, exact).  UI environment is the
parsed function for RxODE.  \code{exact} is a boolean that says if an
exact match is required.}

\item{...}{Other arguments}
}
\value{
value that was requested from the UI object
}
\description{
S3 for getting information from UI model
}
\author{
Matthew Fidler
}
