% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxUnload}
\alias{rxUnload}
\title{Unload the DLL for the object}
\usage{
rxUnload(obj)
}
\arguments{
\item{obj}{a RxODE family of objects}
}
\description{
This unloads the DLL in the R session so that the DLL can be
deleted.  All the c functions will no longer be accessible.
}
\author{
Matthew L.Fidler
}
