% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxUnloadAll}
\alias{rxUnloadAll}
\title{Unloads all RxODE compiled DLLs}
\usage{
rxUnloadAll()
}
\value{
List of RxODE dlls still loaded

boolean of if all RxODE dlls have been unloaded
}
\description{
Unloads all RxODE compiled DLLs
}
\examples{

print(rxUnloadAll())

}
