% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utils.R
\name{rxUnloadAll_}
\alias{rxUnloadAll_}
\alias{rxUnloadAll}
\title{Unload all RxODE Dlls that are not locked for solving.}
\usage{
rxUnloadAll_()

rxUnloadAll()
}
\description{
Unload all RxODE Dlls that are not locked for solving.
}
