% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxUseRadixSort}
\alias{rxUseRadixSort}
\title{Use Radix Sort when possible}
\usage{
rxUseRadixSort(useRadix = TRUE)
}
\arguments{
\item{useRadix}{Use \code{order} with \code{method} = \code{radix}
when appropriate.  Otherwise use \code{timsort}.}
}
\description{
By default RxODE uses radix sort when possible.
}
\author{
Matthew Fidler
}
