% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxValidate.R
\name{rxValidate}
\alias{rxValidate}
\alias{rxTest}
\title{Validate RxODE}
\usage{
rxValidate(type = NULL, check = FALSE)

rxTest(type = NULL, check = FALSE)
}
\arguments{
\item{type}{Type of test or filter of test type}

\item{check}{Use devtools::check instead}
}
\description{
This allows easy validation/qualification of nlmixr by running the
testing suite on your system.
}
\author{
Matthew L. Fidler
}
