% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxLogicToVar.R
\name{rxVarToLogic}
\alias{rxVarToLogic}
\title{Convert logical variable list and model to logical model}
\usage{
rxVarToLogic(variables, model)
}
\arguments{
\item{variables}{Logical variables that will replace the rx_lgl_#
in the model}

\item{model}{A model with rx_lgl_# information contained within}
}
\value{
Model with the rx_lgl_# variables replaced with the
    logical variables in the data
}
\description{
Convert logical variable list and model to logical model
}
\author{
Matthew Fidler
}
\keyword{internal}
