% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{rxVersion}
\alias{rxVersion}
\title{Version and repository for this RxODE package.}
\usage{
rxVersion()
}
\value{
A character vector with the version and repository.
}
\description{
Version and repository for this RxODE package.
}
\author{
Matthew L. Fidler
}
\keyword{internal}

