% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE-win-setup.R
\name{rxWinSetup}
\alias{rxWinSetup}
\title{Setup Windows components for RxODE}
\usage{
rxWinSetup(rm.rtools = TRUE)
}
\arguments{
\item{rm.rtools}{Remove Rtools from path?}
}
\description{
Setup Windows components for RxODE
}
\author{
Matthew L. Fidler
}
