% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{sample_frac}
\alias{sample_frac}
\alias{sample_frac.solveRxDll}
\title{sample_frac for \code{solveRxDll} object}
\usage{
\method{sample_frac}{solveRxDll}(tbl, ...)
}
\arguments{
\item{tbl}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
compatability function for dplyr
}
