% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{sample_n.solveRxDll}
\alias{sample_frac.solveRxDll}
\alias{sample_n.solveRxDll}
\title{dyplr support of sample_n and sample_frac}
\usage{
sample_n.solveRxDll(tbl, size, replace = FALSE, weight = NULL,
  .env = parent.frame())

sample_frac.solveRxDll(tbl, size = 1, replace = FALSE, weight = NULL,
  .env = parent.frame())
}
\arguments{
\item{tbl}{solved RxODE object}

\item{size}{size of sampled object}

\item{replace}{Sample with or without replacement?}

\item{weight}{Sampling weights. This expression is evaluated in the
context of the data frame. It must return a vector of
non-negative numbers the same length as the input. Weights are
automatically standardised to sum to 1.}

\item{.env}{Environment in which to look for non-data names used in
\code{weight}. Non-default settings for experts only.}
}
\description{
sample_n samples n from the solved RxODE solved object.
}
\details{
sample_frac samples a fraction of the number rows of the solved
RxODE object.
}
\keyword{internal}

