% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{select_}
\alias{select_}
\alias{select_.solveRxDll}
\title{select_ for \code{solveRxDll} object}
\usage{
\method{select_}{solveRxDll}(data, ...)
}
\arguments{
\item{data}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
