% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{select_.solveRxDll}
\alias{select_.solveRxDll}
\title{dyplr select_ support}
\usage{
select_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that select_ is being applied to.}

\item{...}{arguments to select_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr select_ support
}
\keyword{internal}

