% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{separate_}
\alias{separate_}
\alias{separate_.solveRxDll}
\title{separate_ for \code{solveRxDll} object}
\usage{
\method{separate_}{solveRxDll}(data, ...)
}
\arguments{
\item{data}{Solved ODE, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
compatability function for tidyr
}
