% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{separate_.solveRxDll}
\alias{separate_.solveRxDll}
\title{tidyr separate support}
\usage{
separate_.solveRxDll(data, col, into, sep = "[^[:alnum:]]+", remove = TRUE,
  convert = FALSE, extra = "warn", fill = "warn", ...)
}
\arguments{
\item{data}{RxODE solved object}

\item{col}{name of column to split, as string}

\item{into}{Names of new variables to create as character vector.}

\item{sep}{Separator between columns.  See tidyr for more
information.}

\item{remove}{If \code{TRUE}, remove input column from output data
frame.}

\item{convert}{If 'TRUE', will run 'type.convert' with 'as.is =
TRUE' on new columns.}

\item{extra}{If \code{sep} is a character vector, this controls
what happens when there are too many pieces.  See tidyr
separate for more details.}

\item{fill}{If \code{sep} is a character vector, this controls what
happens when there are not enough pieces.  See tidyr separate
for more details.}

\item{...}{Ignored}
}
\description{
tidyr separate support
}
\keyword{internal}

