% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{slice_}
\alias{slice_}
\alias{slice_.solveRxDll}
\title{slice_ for \code{solveRxDll} object}
\usage{
\method{slice_}{solveRxDll}(.data, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{data}{Solved equation, an \code{solveRxDll} object.}
}
