% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{slice_.solveRxDll}
\alias{slice_.solveRxDll}
\title{dyplr slice_ support}
\usage{
slice_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that slice_ is being applied to.}

\item{...}{arguments to slice_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr slice_ support
}
\keyword{internal}

