% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{solve.rxDll}
\alias{solve.rxDll}
\title{Solves a rxDll object}
\usage{
\method{solve}{rxDll}(rxDllObj, params, events, inits = NULL, stiff = TRUE,
  transit_abs = FALSE, atol = 1e-08, rtol = 1e-06, ...)
}
\arguments{
\item{rxDllObj}{is a rxDll object}

\item{params}{is a named vector of parameter estimates.}

\item{events}{is an eventTable object of events and sampling times}

\item{inits}{is a named vector of the inital states of the system.}

\item{stiff}{is a boolean indicating if  the system stiff}

\item{atol}{absolute tolerance for LSODA}

\item{rtol}{relative tolerance for LSODA}

\item{tranit_abs}{boolean indicating if this is a transit compartment absorption}
}
\description{
This uses a rxDll object to solve a system.
}
\seealso{
\code{\link{RxODE}}
}
\keyword{internal}

