% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{spread_}
\alias{spread_}
\alias{spread_.solveRxDll}
\title{spread_ for \code{solveRxDll} object}
\usage{
\method{spread_}{solveRxDll}(data, ...)
}
\arguments{
\item{data}{Solved ODE, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
