% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{spread_.solveRxDll}
\alias{spread_.solveRxDll}
\title{tidyr's spread_ support}
\usage{
spread_.solveRxDll(data, key_col, value_col, fill = NA, convert = FALSE,
  drop = TRUE)
}
\arguments{
\item{data}{RxODE solved data.}

\item{key_col, value_col}{Strings giving names of key and value cols.}

\item{fill}{If set, missing values will be replaced with this
value. See spread for more details.}

\item{convert}{If \code{TRUE}, \code{type.convert} with \code{asis
= TRUE} will be run on each of the new columns.  See spread for
more details.}

\item{drop}{If \code{FALSE}, will keep factor levels that don't
appear in the data, filling in missing combinations with
\code{fill}.}
}
\description{
tidyr's spread_ support
}
\keyword{internal}

