% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{[[<-.solveRxDll}
\alias{[[<-.solveRxDll}
\title{Assign solved objects using the [[]] syntax}
\usage{
\method{[[}{solveRxDll}(obj, arg) <- value
}
\arguments{
\item{obj}{solved object}

\item{arg}{element of solved object}

\item{value}{value assumed}
}
\description{
Assign solved objects using the [[]] syntax
}
\seealso{
\code{\link{rxSolve}}
}
\keyword{internal}

