% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{[<-.solveRxDll}
\alias{[<-.solveRxDll}
\title{Assign solved objects using the [] syntax}
\usage{
\method{[}{solveRxDll}(obj, arg1, arg2) <- value
}
\arguments{
\item{obj}{solved object}

\item{arg1}{first argument}

\item{arg2}{second argument}

\item{value}{value assumed}
}
\description{
Assign solved objects using the [] syntax
}
\keyword{internal}

