% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{summarise_}
\alias{summarise_}
\alias{summarise_.solveRxDll}
\title{summarise_ for \code{solveRxDll} object}
\usage{
\method{summarise_}{solveRxDll}(data, ...)
}
\arguments{
\item{data}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
