% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{summarise_.solveRxDll}
\alias{summarise_.solveRxDll}
\title{dyplr summarise_ support}
\usage{
\method{summarise_}{solveRxDll}(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that summarise_ is being applied to.}

\item{...}{arguments to summarise_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr summarise_ support
}
\author{
Matthew L.Fidler
}
\keyword{internal}

