% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{summary.rxDll}
\alias{summary.rxDll}
\title{Summary of rxDll object}
\usage{
\method{summary}{rxDll}(x, ..., noprint = FALSE)
}
\arguments{
\item{noprint}{is a logical telling if it should print the rxDll
object fist.}
}
\description{
This gives expanded information about the rxDll object
}
\keyword{internal}

