% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{summary.RxODE}
\alias{summary.RxODE}
\title{Print expanded information about the RxODE object.}
\usage{
\method{summary}{RxODE}(x, ...)
}
\arguments{
\item{...}{Ignored parameters}

\item{RxODEobj}{RxODE object}
}
\description{
This prints the expanded information about the RxODE object.
}

