% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{tail.solveRxDll}
\alias{tail.solveRxDll}
\title{Return the tail of RxODE solved data}
\usage{
\method{tail}{solveRxDll}(x, n = 6L, addrownums = TRUE, ...)
}
\description{
Extract the matrix and return the tail of the matrix
}

