% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{transmute_}
\alias{transmute_}
\alias{transmute_.solveRxDll}
\title{transmute_ for \code{solveRxDll} object}
\usage{
\method{transmute_}{solveRxDll}(.data, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{data}{Solved equation, an \code{solveRxDll} object.}
}
