% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{transmute_.solveRxDll}
\alias{transmute_.solveRxDll}
\title{dyplr transmute_ support}
\usage{
transmute_.solveRxDll(.data, ..., .dots)
}
\arguments{
\item{.data}{solveRxDll object that transmute_ is being applied to.}

\item{...}{arguments to transmute_ method}

\item{.dots}{dplyr .dots argument}
}
\description{
dyplr transmute_ support
}
\keyword{internal}

