% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{unite_}
\alias{unite_}
\alias{unite_.solveRxDll}
\title{unite_ for \code{solveRxDll} object}
\usage{
\method{unite_}{solveRxDll}(data, ...)
}
\arguments{
\item{data}{Solved ODE, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
