% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{unite_.solveRxDll}
\alias{unite_.solveRxDll}
\title{tidyr unite support}
\usage{
unite_.solveRxDll(data, col, from, sep = "_", remove = TRUE)
}
\arguments{
\item{data}{solved RxODE object}

\item{col}{Name of new column as string}

\item{from}{Names of existing columns as character vector}

\item{sep}{Separator to use between values}

\item{remove}{If \code{TRUE}, remove input columns from output data
frame.}
}
\description{
tidyr unite support
}
\keyword{internal}

