% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{update.solveRxDll}
\alias{update.solveRxDll}
\title{Update the solved object with any of the new parameters.}
\usage{
\method{update}{solveRxDll}(object, ...)
}
\arguments{
\item{object}{Object to be updated}

\item{...}{Arguments to be updated, and resolved.}
}
\description{
This is a wrapper to the rxSolve method.
}
\author{
Matthew L.Fidler
}
