% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{update.solveRxODE}
\alias{update.solveRxODE}
\title{Update the solved object with any of the new parameters.}
\usage{
\method{update}{solveRxODE}(object, ...)
}
\arguments{
\item{object}{Object to be updated}

\item{...}{Arguments to be updated, and resolved.}
}
\description{
Update the solved object with any of the new parameters.
}
\author{
Matthew L.Fidler
}
