% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uppergamma}
\alias{uppergamma}
\title{uppergamma:  upper incomplete gamma function}
\usage{
uppergamma(a, z)
}
\arguments{
\item{a}{The numeric 'a' parameter in the upper
incomplete gamma}

\item{z}{The numeric 'z' parameter in the upper
incomplete gamma}
}
\value{
uppergamma results
}
\description{
This is the tgamma from the boost library
}
\details{
The uppergamma function is given by:

\eqn{uppergamma(a, z) = \int_{z}^{\infty}t^{a-1}\cdot e^{-t} dt}
}
\examples{

uppergamma(1,3)

uppergamma(1:3,3)

uppergamma(1,1:3)

}
\author{
Matthew L. Fidler
}
