% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{as.tbl.solveRxDll}
\alias{as.tbl.solveRxDll}
\title{Convert Solved RxODE object to tbl}
\usage{
as.tbl.solveRxDll(x, ...)
}
\arguments{
\item{x}{Solved RxDll object}

\item{...}{other arguments (ignored)}
}
\description{
Convert Solved RxODE object to tbl
}

