% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{as_data_frame}
\alias{as_data_frame}
\alias{as_data_frame.solveRxODE}
\alias{as_data_frame}
\alias{as.tbl.solveRxODE}
\title{as_data_frame for solveRxODE}
\usage{
\method{as_data_frame}{solveRxODE}(x, ...)

\method{as_data_frame}{solveRxODE}(x, ...)
}
\arguments{
\item{x}{Solved data frame}

\item{...}{Additional arguments}

\item{x}{Solved data frame}

\item{...}{Additional arguments}
}
\description{
compatability function for as_data_frame

compatability function for as_data_frame
}
\author{
Matthew L. Fidler

Matthew L. Fidler
}
\keyword{internal}
