% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{as_tibble.rxEt}
\alias{as_tibble.rxEt}
\alias{as.tbl.rxEt}
\title{Convert to tbl}
\usage{
as_tibble.rxEt(x, ...)

as.tbl.rxEt(x, ...)
}
\arguments{
\item{x}{RxODE event table}

\item{...}{Other arguments to \code{as.tbl}}
}
\value{
tibble
}
\description{
Convert to tbl
}
