% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rbind.solveRxDll}
\alias{cbind.solveRxDll}
\alias{rbind.solveRxDll}
\title{cbind/rbind solveRxDll}
\usage{
rbind.solveRxDll(...)

cbind.solveRxDll(...)
}
\arguments{
\item{...}{ignored parameters}
}
\description{
Cbind/rbind is disabled for RxOde solved objects.  Use as.data.frame(x)
to use cbind/rbind.
}
\keyword{internal}

