% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvPost}
\alias{cvPost}
\title{Sample A covariance Matrix.}
\usage{
cvPost(nu, Omega, omegaIsChol = FALSE)
}
\arguments{
\item{omegaIsChol}{is an indicator of if the Omega matrix is in the cholesky decomposition.}
}
\description{
Sample A covariance Matrix.
}
\author{
Matthew L.Fidler & Wenping Wang
}
