% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenEtaS}
\alias{.rxGenEtaS}
\title{Generate the ETA sensitivities for FO related methods}
\usage{
.rxGenEtaS(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL)
}
\value{
RxODE/symengine environment
}
\description{
Generate the ETA sensitivities for FO related methods
}
\author{
Matthew L. Fidler
}
