% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenFun}
\alias{.rxGenFun}
\title{Generate and load RxODE function into sympy environment}
\usage{
.rxGenFun(
  obj,
  predfn,
  pkpars = NULL,
  errfn = NULL,
  init = NULL,
  promoteLinSens = TRUE,
  full = TRUE
)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}
}
\value{
Sympy environment
}
\description{
Generate and load RxODE function into sympy environment
}
\author{
Matthew Fidler
}
