% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenPkpars}
\alias{.rxGenPkpars}
\title{Generate a RxODE block augmented by a pkpars function}
\usage{
.rxGenPkpars(rx, pkpars, init)
}
\value{
New model augmented by pred function
}
\description{
Generate a RxODE block augmented by a pkpars function
}
\author{
Matthew L. Fidler
}
