% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{.rxGenPred}
\alias{.rxGenPred}
\title{Generate Augmented pred/err RxODE model}
\usage{
.rxGenPred(obj, predfn, errfn, init)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}
}
\value{
A list of (1) RxODE model variables augmented with
    pred/error information and (2) extra error variables created.
}
\description{
Generate Augmented pred/err RxODE model
}
\author{
Matthew L Fidler
}
