% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE-win-setup.R
\name{.rxWinRtoolsPath}
\alias{.rxWinRtoolsPath}
\title{Setup Rtools path}
\usage{
.rxWinRtoolsPath(rm.rtools = TRUE, rm.python = TRUE)
}
\arguments{
\item{rm.rtools}{Remove the Rtools from the current path specs.}

\item{rm.python}{Remove Python from the current path specs.}
}
\description{
Setup Rtools path
}
\author{
Matthew L. Fidler
}
