% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rxWithOptions}
\alias{.rxWithOptions}
\title{Temporarily set options then restore them while running code}
\usage{
.rxWithOptions(ops, code)
}
\arguments{
\item{ops}{list of options that will be temporarily set for the
\code{code}}

\item{code}{The code to run during the sink}
}
\value{
value of code
}
\description{
Temporarily set options then restore them while running code
}
\examples{

.rxWithOptions(list(digits = 21), {
  print(pi)
})

print(pi)
}
