% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etRbind}
\alias{etRbind}
\title{Combining event tables}
\usage{
etRbind(..., samples = c("use", "clear"), waitII = c("smart", "+ii"),
  id = c("merge", "unique"))
}
\arguments{
\item{samples}{can be "clear", "use"}
}
\value{
A new event table
}
\description{
Combining event tables
}
