% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etRbind}
\alias{etRbind}
\alias{rbind.rxEt}
\title{Combining event tables}
\usage{
etRbind(..., samples = c("use", "clear"), waitII = c("smart", "+ii"),
  id = c("merge", "unique"))

\method{rbind}{rxEt}(..., deparse.level = 1)
}
\arguments{
\item{...}{The event tables and optionally time between event
tables, called waiting times in this help document.}

\item{samples}{How to handle samples when repeating an event
    table.  The options are:
\itemize{

\item{"clear"} Clear
    sampling records before combining the datasets

\item{"use"}
    Use the sampling records when combining the datasets

}}

\item{waitII}{This determines how waiting times between events are
    handled. The options are:

\itemize{

\item \code{"smart"} This "smart" handling of waiting times is the
default option.  In this case, if the waiting time is above the
last observed inter-dose interval in the first combined event
table, then the actual time between doses is given by the wait
time.  If it is smaller than the last observed inter-dose
interval, the time between event tables is given by the inter-dose
interval + the waiting time between event tables.

\item \code{"+ii"} In this case, the wait time is added to the
inter-dose interval no matter the length of the wait time or
inter-dose interval

}}

\item{id}{This is how rbind will handle IDs.  There are two different types of options:
\itemize{

\item{merge} with \code{id="merge"}, the IDs are merged together,
overlapping IDs would be merged into a single event table.

\item{unique} with \code{id="unique"}, the IDs will be renumbered
so that the IDs in all the event tables are not overlapping.

}}

\item{deparse.level}{The \code{deparse.level} of a traditional
\code{rbind} is ignored.}
}
\description{
Combining event tables
}
\examples{

}
\author{
Matthew L Fidler
}
