% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etRep}
\alias{etRep}
\title{Repeat an RxODE event table}
\usage{
etRep(x, times = 1, length.out = NA, each = NA, n = NULL,
  wait = 0, id = integer(0), samples = c("clear", "use"),
  waitII = c("smart", "+ii"), ii = 24)
}
\arguments{
\item{times}{Number of times to repeat the event table}

\item{length.out}{Invalid with RxODE event tables, will throw an
error if used.}

\item{each}{Invalid with RxODE event tables, will throw an error
if used.}

\item{n}{The number of times to repeat the event table.  Overrides
\code{times}.}

\item{wait}{Waiting time between each repeated event table.  By
default there is no waiting, or wait=0}

\item{id}{IDs to expand/remove in the event table before repeating.}

\item{samples}{can be "clear", "use"}
}
\value{
An event table of repeated events
}
\description{
Repeat an RxODE event table
}
\author{
Matthew Fidler
}
