% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{etSeq}
\alias{etSeq}
\title{.. content for \description{} (no empty lines) ..}
\usage{
etSeq(..., handleSamples = c("clear", "use"))
}
\arguments{
\item{handleSamples}{is a boolean indicating if the sampling
should be cleared when combining the event tables.}
}
\description{
.. content for \description{} (no empty lines) ..
}
