% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{etTrans}
\alias{etTrans}
\title{Event translation for RxODE}
\usage{
etTrans(inData, obj, addCmt = FALSE)
}
\arguments{
\item{inData}{Data frame to translate}

\item{obj}{Model to translate data}

\item{addCmt}{Add compartment to data frame, and drop units}
}
\value{
Object for solving in RxODE
}
\description{
Event translation for RxODE
}
\keyword{internal}
