% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{guide_none}
\alias{guide_none}
\title{Empty Guide}
\usage{
guide_none(title = waiver(), position = waiver())
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\value{
nothing, simply included to be compatible with ggplot 3.2
}
\description{
This empty guide draws nothing; It is included in RxODE for
compatibility with ggplot 3.2
}
