% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping-ini.R, R/ui.R
\name{ini.rxUi}
\alias{ini.rxUi}
\alias{ini.function}
\alias{ini}
\alias{ini.default}
\title{Ini block for RxODE/nlmixr models}
\usage{
\method{ini}{rxUi}(x, ..., envir = parent.frame())

\method{ini}{`function`}(x, ..., envir = parent.frame())

ini(x, ..., envir = parent.frame())

\method{ini}{default}(x, ...)
}
\arguments{
\item{x}{expression}

\item{...}{Other expressions for \code{ini()} function}

\item{envir}{the \code{environment} in which unevaluated model
expressions is to be evaluated.  May also be \code{NULL}, a list, a
data frame, a pairlist or an integer as specified to \code{sys.call}.}
}
\value{
Ini block
}
\description{
Ini block for RxODE/nlmixr models
}
\author{
Matthew Fidler
}
