% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ini}
\alias{ini}
\title{Ini block for RxODE/nlmixr models}
\usage{
ini(x, ...)
}
\arguments{
\item{x}{expression}
}
\value{
Ini block
}
\description{
Ini block for RxODE/nlmixr models
}
\author{
Matthew Fidler
}
