% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping-model.R, R/ui.R
\name{model.function}
\alias{model.function}
\alias{model.rxUi}
\alias{model}
\alias{model.default}
\title{Model block for RxODE/nlmixr models}
\usage{
\method{model}{`function`}(x, ..., envir = parent.frame())

\method{model}{rxUi}(x, ..., envir = parent.frame())

model(x, ..., envir = parent.frame())

\method{model}{default}(x, ...)
}
\arguments{
\item{x}{model expression}

\item{...}{Other arguments}

\item{envir}{the \code{environment} in which unevaluated model
expressions is to be evaluated.  May also be \code{NULL}, a list, a
data frame, a pairlist or an integer as specified to \code{sys.call}.}
}
\value{
Model block with ini information included.  \code{ini} must be
called before \code{model} block
}
\description{
Model block for RxODE/nlmixr models
}
\author{
Matthew Fidler
}
