% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{plot.RxODE}
\alias{plot.RxCompilationManager}
\alias{plot.RxODE}
\alias{plot.rxDll}
\title{Plot a model digram for simple ODE models}
\usage{
\method{plot}{RxODE}(x, ...)

\method{plot}{rxDll}(x, ...)

\method{plot}{RxCompilationManager}(x, ...)
}
\arguments{
\item{x}{is an RxODE object}

\item{...}{are other arguments

This function is an alias for the \code{\link{rxPlot}} function.}
}
\description{
This plots a line diagram for the current system of differential
equations.
}
\seealso{
\code{\link{rxPlot}},\code{\link{RxODE}}.
}

