% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rxCoef}
\alias{print.rxCoef}
\title{Print the rxCoef object}
\usage{
\method{print}{rxCoef}(x, ...)
}
\arguments{
\item{x}{rxCoef object}
}
\value{
original object
}
\description{
This prints out the user supplied arguments for rxCoef object
}
\author{
Matthew L.Fidler
}
\keyword{internal}
