% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.RxODE}
\alias{print.RxODE}
\title{Print information about the RxODE object.}
\usage{
\method{print}{RxODE}(x, ...)
}
\arguments{
\item{x}{An rxode object}

\item{...}{Ignored parameters}
}
\value{
original object
}
\description{
This prints the model name and its status for being able to be solved
}
\author{
Matthew L.Fidler
}
