% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{aes}
\alias{facet_wrap}
\alias{geom_line}
\alias{ggplot}
\alias{scale_type}
\alias{set_units}
\alias{drop_units}
\alias{ylab}
\alias{xlab}
\alias{waiver}
\alias{lotri}
\alias{type_sum}
\alias{label_value}
\alias{label_both}
\alias{label_context}
\alias{label_wrap_gen}
\alias{scale_x_discrete}
\alias{scale_y_discrete}
\alias{scale_x_continuous}
\alias{scale_y_continuous}
\alias{scale_x_date}
\alias{scale_y_date}
\alias{expand_limits}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{expand_limits}}, \code{\link[ggplot2]{facet_wrap}}, \code{\link[ggplot2:geom_path]{geom_line}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2:labellers]{label_both}}, \code{\link[ggplot2:labellers]{label_context}}, \code{\link[ggplot2:labellers]{label_context}}, \code{\link[ggplot2:labellers]{label_value}}, \code{\link[ggplot2:labellers]{label_wrap_gen}}, \code{\link[ggplot2]{scale_type}}, \code{\link[ggplot2:scale_continuous]{scale_x_continuous}}, \code{\link[ggplot2:scale_date]{scale_x_date}}, \code{\link[ggplot2:scale_discrete]{scale_x_discrete}}, \code{\link[ggplot2:scale_continuous]{scale_y_continuous}}, \code{\link[ggplot2:scale_date]{scale_y_date}}, \code{\link[ggplot2:scale_discrete]{scale_y_discrete}}, \code{\link[ggplot2]{waiver}}, \code{\link[ggplot2:labs]{xlab}}, \code{\link[ggplot2:labs]{ylab}}}

  \item{lotri}{\code{\link[lotri]{lotri}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{pillar}{\code{\link[pillar]{type_sum}}}

  \item{units}{\code{\link[units]{drop_units}}, \code{\link[units]{set_units}}}
}}

