% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvchisq}
\alias{rinvchisq}
\title{Scaled Inverse Chi Squared distribution}
\usage{
rinvchisq(n = 1L, nu = 1, scale = 1)
}
\arguments{
\item{n}{Number of random samples}

\item{nu}{degrees of freedom of inverse chi square}

\item{scale}{Scale of inverse chi squared distribution 
(default is 1).}
}
\value{
a vector of inverse chi squared deviates .
}
\description{
Scaled Inverse Chi Squared distribution
}
