% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxChain2}
\alias{rxChain2}
\alias{rxChain2.default}
\alias{rxChain2.EventTable}
\title{Second command in chaining commands}
\usage{
rxChain2(obj, solvedObject)

\method{rxChain2}{default}(obj, solvedObject)

\method{rxChain2}{EventTable}(obj, solvedObject)
}
\arguments{
\item{obj}{the object being added/chained/piped to the solved object}

\item{solvedObject}{the solved object}
}
\value{
chained operation
}
\description{
This is s3 method is called internally with \code{+} and \verb{\\\%>\\\%} operators.
}
\author{
Matthew L.Fidler
}
\keyword{internal}
