% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxClean}
\alias{rxClean}
\title{Cleanup anonymous DLLs}
\usage{
rxClean(wd)
}
\arguments{
\item{wd}{What directory should be cleaned

This cleans up all files named \code{rx-*.dll} and associated files as
well as \code{call_dvode.o} and associated files}
}
\value{
TRUE if successful
}
\description{
This cleans up any DLLs created by text files
}
\author{
Matthew L. Fidler
}
