% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxClean}
\alias{rxClean}
\title{Cleanup anonymous DLLs}
\usage{
rxClean(wd)
}
\arguments{
\item{wd}{What directory should be cleand

This cleans up all files named rx-*.dll and associated files as
well as call_dvode.o and associated files}
}
\value{
TRUE if successful
}
\description{
This cleans up any DLLs created by text files
}
\author{
Matthew L. Fidler
}
