% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDataParSetup}
\alias{rxDataParSetup}
\title{Setup Data and Parameters}
\usage{
rxDataParSetup(object, params = NULL, events = NULL, inits = NULL,
  covs = NULL, sigma = NULL, sigmaDf = NULL, nCoresRV = 1L,
  sigmaIsChol = FALSE, nDisplayProgress = 10000L,
  amountUnits = NA_character_, timeUnits = "hours", theta = NULL,
  eta = NULL, scale = NULL, extraArgs = NULL)
}
\arguments{
\item{object}{is a either a RxODE family of objects, or a file-name
with a RxODE model specification, or a string with a RxODE
model specification.}

\item{params}{a numeric named vector with values for every
parameter in the ODE system; the names must correspond to the
parameter identifiers used in the ODE specification;}

\item{events}{an \code{eventTable} object describing the input
(e.g., doses) to the dynamic system and observation sampling
time points (see \code{\link{eventTable}});}

\item{inits}{a vector of initial values of the state variables
(e.g., amounts in each compartment), and the order in this
vector must be the same as the state variables (e.g., PK/PD
compartments);}

\item{covs}{a matrix or dataframe the same number of rows as the
sampling points defined in the events \code{eventTable}.  This
is for time-varying covariates.}

\item{sigma}{Named sigma matrix.}

\item{sigmaDf}{The degrees of freedom of a t-distribution for
simulation.  By default this is \code{NULL} which is
equivalent to \code{Inf} degrees, or to simulate from a normal
distribution instead of a t-distribution.}

\item{nCoresRV}{Number of cores for residual simulation.  This,
along with the seed, affects both the outcome and speed of
simulation. By default it is one.}

\item{sigmaIsChol}{Indicates if the \code{sigma} supplied is a
Cholesky decomposed matrix instead of the traditional
symmetric matrix.}

\item{nDisplayProgress}{An integer indicating the minimum number
of c-based solves before a progress bar is shown.  By default
this is 10,000.}

\item{amountUnits}{This supplies the dose units of a data frame
supplied instead of an event table.  This is for importing the
data as an RxODE event table.}

\item{timeUnits}{This supplies the time units of a data frame
supplied instead of an event table.  This is for importing the
data as an RxODE event table.}

\item{theta}{A vector of parameters that will be named THETA[#] and
added to parameters}

\item{eta}{A vector of parameters that will be named ETA[#] and
added to parameters}

\item{scale}{a numeric named vector with scaling for ode
parameters of the system.  The names must correstond to the
parameter identifiers in the ODE specification. Each of the
ODE variables will be divided by the scaling factor.  For
example \code{scale=(center=2)} will divide the center ODE
variable by 2.}
}
\value{
Data setup for running C-based RxODE runs.
}
\description{
Setup Data and Parameters
}
\author{
Matthew L. Fidler
}
\keyword{internal}
