% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve.R
\name{rxDataParSetup}
\alias{rxDataParSetup}
\title{Setup Data and Parameters}
\usage{
rxDataParSetup(object, params = NULL, events = NULL, inits = NULL,
  sigma = NULL, sigma.df = Inf, sigma.ncores = 1L, sigma.isChol = FALSE)
}
\arguments{
\item{object}{is a either a RxODE family of objects, or a file-name
with a RxODE model specification, or a string with a RxODE
model specification.}

\item{params}{a numeric named vector with values for every
parameter in the ODE system; the names must correspond to the
parameter identifiers used in the ODE specification;}

\item{events}{an \code{eventTable} object describing the input
(e.g., doses) to the dynamic system and observation sampling
time points (see \code{\link{eventTable}});}

\item{inits}{a vector of initial values of the state variables
(e.g., amounts in each compartment), and the order in this
vector must be the same as the state variables (e.g., PK/PD
compartments);}

\item{sigma}{Named sigma matrix.}

\item{sigma.df}{The degrees of freedom of a t-distribution for
simulation.  By default this is \code{Inf}, or to simulate
from a normal distribution instead of a t.}

\item{sigma.ncores}{Number of cores for residual simulation.
This, along with the seed, affects both the outcome and speed
of simulation. By default it is one.}

\item{sigma.isChol}{Indicates if the \code{sigma} supplied is a
Cholesky decomposed matrix instead of the traditional
symmetric matrix.}
}
\value{
Data setup for running C-based RxODE runs.r
}
\description{
Setup Data and Parameters
}
\author{
Matthew L. Fidler
}
