% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDataSetup}
\alias{rxDataSetup}
\title{Setup a data frame for solving multiple subjects at once in RxODE.}
\usage{
rxDataSetup(ro, covNames = NULL, amountUnits = NA_character_,
  timeUnits = "hours")
}
\arguments{
\item{ro}{R object to setup; Must be in RxODE compatible format.}

\item{covNames}{Covariate names in dataset.}

\item{amountUnits}{Dosing amount units.}

\item{timeUnits}{Time units.}
}
\value{
A data structure to allow C-based for loop (ie solving each
      individual in C)
}
\description{
Setup a data frame for solving multiple subjects at once in RxODE.
}
