% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDataSetup}
\alias{rxDataSetup}
\title{Setup a data frame for solving multiple subjects at once in RxODE.}
\usage{
rxDataSetup(ro, covNames = NULL, sigma = NULL, df = NULL, ncoresRV = 1L,
  isChol = FALSE, amountUnits = NA_character_, timeUnits = "hours")
}
\arguments{
\item{ro}{R object to setup; Must be in RxODE compatible format.}

\item{covNames}{Covariate names in dataset.}

\item{sigma}{Matrix for simulating residual variability for the number of observations.
     This uses the \code{\link[mvnfast]{rmvn}} or \code{\link[mvnfast]{rmvt}} assuming mean 0 and covariance given by this named
     matrix.  The residual-type variability is created for each of the named \"err\" components specified by the sigma matrix's
     column names. This then creates a random deviate that is used in the place of each named variable.  This should
     not really be used in the ODE specification.  If it is, though, it is treated as a time-varying covariate.

     Also note this does not use R's random numbers, rather it uses a cryptographic parallel random number generator;

     To allow reproducible research you must both set a random seed with R's \code{\link[base]{set.seed}} function, and keep the
     number of cores constant.  By changing either one of these variables, you will arrive at different random numbers.}

\item{df}{Degrees of freedom for \code{\link[mvnfast]{rmvt}}.  If \code{NULL}, or \code{Inf}, then use a normal distribution.  The
default is normal.}

\item{ncoresRV}{The number of cores for residual simulation.  By default, this is \code{1}.}

\item{isChol}{is a boolean indicating that the Cholesky decomposition of the \code{sigma} covariance matrix is supplied instead of
the  \code{sigma} matrix itself.}

\item{amountUnits}{Dosing amount units.}

\item{timeUnits}{Time units.}
}
\value{
A data structure to allow C-based for loop (ie solving each
      individual in C)
}
\description{
Setup a data frame for solving multiple subjects at once in RxODE.
}
\author{
Matthew L. Fidler
}
