% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDelete}
\alias{rxDelete}
\title{Delete the DLL for the model}
\usage{
rxDelete(obj)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
A boolean stating if the operation was successful.
}
\description{
This function deletes the DLL, but doesn't delete the model
information in the object.
}
\author{
Matthew L.Fidler
}
