% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxDfdy}
\alias{rxDfdy}
\title{Jacobain and parameter derivates}
\usage{
rxDfdy(obj)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
A list of the jacobian parameters defined in this RxODE
    object.
}
\description{
Return Jacobain and parameter derivates
}
\author{
Matthew L. Fidler
}
