% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDfdy}
\alias{rxDfdy}
\title{Jacobian and parameter derivatives}
\usage{
rxDfdy(obj)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
A list of the jacobian parameters defined in this RxODE
object.
}
\description{
Return Jacobain and parameter derivatives
}
\author{
Matthew L. Fidler
}
