% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxDll}
\alias{rxDll}
\title{Return the DLL associated with the RxODE object}
\usage{
rxDll(obj)
}
\arguments{
\item{obj}{A RxODE family of objects or a character string of the
model specification or location of a file with a model
specification.}
}
\value{
a path of the library
}
\description{
This will return the dynamic load library or shared object used to
run the C code for RxODE.
}
\author{
Matthew L.Fidler
}
\keyword{internal}
