% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxDynProtect}
\alias{rxDynProtect}
\title{Protect dll(s) from unloading}
\usage{
rxDynProtect(dlls)
}
\arguments{
\item{dlls}{is a list of dlls that shouldn't be unloaded}
}
\description{
Protect dll(s) from unloading
}
\author{
Matthew Fidler
}
