% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{rxEvid}
\alias{rxEvid}
\alias{as.rxEvid}
\alias{c.rxEvid}
\alias{[.rxEvid}
\alias{as.character.rxEvid}
\alias{format.rxEvid}
\alias{print.rxEvid}
\alias{[[.rxEvid}
\alias{type_sum.rxEvid}
\alias{pillar_shaft.rxEvid}
\alias{c.rxRateDur}
\alias{format.rxRateDur}
\title{EVID formatting for tibble and other places.}
\usage{
rxEvid(x)

as.rxEvid(x)

\method{c}{rxEvid}(x, ...)

\method{[}{rxEvid}(x, ...)

\method{as.character}{rxEvid}(x, ...)

\method{format}{rxEvid}(x, ...)

\method{print}{rxEvid}(x, ...)

\method{[[}{rxEvid}(x, ...)

\method{type_sum}{rxEvid}(x)

pillar_shaft.rxEvid(x, ...)

\method{c}{rxRateDur}(x, ...)

\method{format}{rxRateDur}(x, ...)
}
\arguments{
\item{x}{Item to be converted to a RxODE EVID specification.}

\item{...}{Other parameters}
}
\description{
This is to make an EVID more readable by non
pharmacometricians. It displays what each means and allows it to
be displayed in a tibble.
}
\examples{

rxEvid(1:7)

}
