% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evTran.R
\name{rxEvTrans}
\alias{rxEvTrans}
\title{Translate Data into solving ready data.}
\usage{
rxEvTrans(data, model)
}
\arguments{
\item{data}{Data to convert}

\item{model}{Model to use for conversion}
}
\value{
Converted data frame
}
\description{
Translate Data into solving ready data.
}
\author{
Matthew Fidler
}
