% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focei.R
\name{rxFoceiInner}
\alias{rxFoceiInner}
\alias{rxFoceiInner.rxFocei}
\alias{rxFoceiInner.RxODE}
\alias{rxFoceiInner.rxDll}
\title{Solve the FOCEI inner problem}
\usage{
rxFoceiInner(object, ..., dv, eta, omegaInv, invisible = 0, m = 6,
  epsilon = 1e-05, past = 0, delta = 0, max_iterations = 0,
  linesearch_algorithm = "LBFGS_LINESEARCH_DEFAULT", max_linesearch = 20,
  min_step = 1e-20, max_step = 1e+20, ftol = 1e-04, wolfe = 0.9,
  gtol = 0.9, orthantwise_c = 0, orthantwise_start = 0,
  orthantwise_end = length(eta))

\method{rxFoceiInner}{rxFocei}(object, ..., dv, eta, omegaInv, invisible = 0,
  m = 6, epsilon = 1e-05, past = 0, delta = 0, max_iterations = 0,
  linesearch_algorithm = "LBFGS_LINESEARCH_DEFAULT", max_linesearch = 20,
  min_step = 1e-20, max_step = 1e+20, ftol = 1e-04, wolfe = 0.9,
  gtol = 0.9, orthantwise_c = 0, orthantwise_start = 0,
  orthantwise_end = length(eta))

\method{rxFoceiInner}{RxODE}(object, ..., dv, eta, omegaInv, invisible = 0,
  m = 6, epsilon = 1e-05, past = 0, delta = 0, max_iterations = 0,
  linesearch_algorithm = "LBFGS_LINESEARCH_DEFAULT", max_linesearch = 20,
  min_step = 1e-20, max_step = 1e+20, ftol = 1e-04, wolfe = 0.9,
  gtol = 0.9, orthantwise_c = 0, orthantwise_start = 0,
  orthantwise_end = length(eta))

\method{rxFoceiInner}{rxDll}(object, ..., dv, eta, omegaInv, invisible = 0,
  m = 6, epsilon = 1e-05, past = 0, delta = 0, max_iterations = 0,
  linesearch_algorithm = "LBFGS_LINESEARCH_DEFAULT", max_linesearch = 20,
  min_step = 1e-20, max_step = 1e+20, ftol = 1e-04, wolfe = 0.9,
  gtol = 0.9, orthantwise_c = 0, orthantwise_start = 0,
  orthantwise_end = length(eta))
}
\arguments{
\item{object}{RxODE object}

\item{...}{values sent to rxFoceiEtaSetup}

\item{dv}{dependent variable}

\item{eta}{eta value}

\item{omegaInv}{inverse omega matrix}

\item{invisible}{Invisible LBFGS tracing.}

\item{m}{LBFGS m}

\item{epsilon}{LBFGS epsilon}

\item{past}{LBFGS past}

\item{delta}{LBFGS delta}

\item{max_iterations}{LBFGS max_iterations}

\item{linesearch_algorithm}{LBFGS linesearch_algorithm}

\item{max_linesearch}{LBFGS max_linesearch}

\item{min_step}{LBFGS min_step}

\item{max_step}{LBFGS max_step}

\item{ftol}{LBFGS ftol}

\item{wolfe}{LBFGS wolfe}

\item{gtol}{LBFGS gtol}

\item{orthantwise_c}{LBFGS orthantwise_c}

\item{orthantwise_start}{LBFGS orthantwise_start}

\item{orthantwise_end}{LBFGS orthantwise_end}
}
\value{
Loglik with fitted and posthoc attributes
}
\description{
Solve the FOCEI inner problem
}
\author{
Matthew L. Fidler
}
\keyword{internal}
