% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{rxGenSaem}
\alias{rxGenSaem}
\title{Generate pred-only SAEM RxODE model}
\usage{
rxGenSaem(obj, predfn, pkpars = NULL, sum.prod = FALSE, optExpression = TRUE)
}
\arguments{
\item{obj}{RxODE model (text or actual model)}

\item{predfn}{prediction function}

\item{pkpars}{PKpars function}

\item{sum.prod}{Change addition and deletion to sum() and prod()
to stabilize round off errors}

\item{optExpression}{Boolean for optimizing expression to minimize
more expensive calls}
}
\value{
RxODE text
}
\description{
Generate pred-only SAEM RxODE model
}
\author{
Matthew Fidler
}
