% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxJacobian.R
\name{rxGenSaem}
\alias{rxGenSaem}
\title{Generate pred-only SAEM RxODE model}
\usage{
rxGenSaem(
  obj,
  predfn,
  pkpars = NULL,
  sum.prod = FALSE,
  optExpression = TRUE,
  loadSymengine = TRUE
)
}
\arguments{
\item{obj}{RxODE model (text or actual model)}

\item{predfn}{prediction function}

\item{pkpars}{PKpars function}

\item{sum.prod}{Change addition and deletion to sum() and prod()
to stabilize round off errors}

\item{optExpression}{Boolean for optimizing expression to minimize
more expensive calls}

\item{loadSymengine}{Boolean indicating if the model should be
loaded into symengine.  This cause all the ODEs to be collapsed
into one expression that is eventually optimized if
\code{optExpression} is \code{TRUE}.}
}
\value{
RxODE text
}
\description{
Generate pred-only SAEM RxODE model
}
\author{
Matthew Fidler
}
