% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxLinCmt.R
\name{rxGetLin}
\alias{rxGetLin}
\title{Get the linear compartment model true function}
\usage{
rxGetLin(model, linCmtSens = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{This is the ODE model specification.  It can be:

\itemize{

 \item a string containing the set of ordinary differential
    equations (ODE) and other expressions defining the changes in
    the dynamic system.

 \item a file name where the ODE system equation is contained

 \item An ODE expression enclosed in \code{\{\}}

  }

(see also the \code{filename} argument). For
    details, see the sections \dQuote{Details} and
    \dQuote{\code{RxODE Syntax}} below.}

\item{linCmtSens}{A boolean indicating if the linCmt() soultions
should be calculated with derivatives (sensitivites) to each
parameter. This is useful for FOCEi and similar calculations.
By default this is \code{FALSE}.}
}
\value{
model with linCmt() replaced with linCmtA()
}
\description{
Get the linear compartment model true function
}
\author{
Matthew Fidler
}
