% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxLinCmt.R
\name{rxGetLin}
\alias{rxGetLin}
\title{Get the linear compartment model true function}
\usage{
rxGetLin(
  model,
  linCmtSens = c("linCmtA", "linCmtB", "linCmtC"),
  verbose = FALSE
)
}
\arguments{
\item{model}{This is the ODE model specification.  It can be:
\itemize{
\item a string containing the set of ordinary differential
equations (ODE) and other expressions defining the changes in
the dynamic system.
\item a file name where the ODE system equation is contained
}

An ODE expression enclosed in \verb{\\\{\\\}}

(see also the \code{filename} argument). For
details, see the sections \dQuote{Details} and
\verb{RxODE Syntax} below.}

\item{linCmtSens}{The method to calculate the linCmt() solutions}

\item{verbose}{When \code{TRUE} be verbose with the linear
compartmental model}
}
\value{
model with linCmt() replaced with linCmtA()
}
\description{
Get the linear compartment model true function
}
\author{
Matthew Fidler
}
