% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxGetModel}
\alias{rxGetModel}
\title{Get model properties without compiling it.}
\usage{
rxGetModel(model, calcSens = NULL, calcJac = NULL,
  collapseModel = NULL)
}
\arguments{
\item{model}{RxODE specification}

\item{calcSens}{boolean indicating if RxODE will calculate the
sennsitivities according to the specified ODEs.}

\item{calcJac}{boolean indicating if RxODE will calculate the
Jacobain according to the specified ODEs.}

\item{collapseModel}{boolean indicating if RxODE will remove all
LHS variables when calculating sensitivities.}
}
\value{
RxODE trans list
}
\description{
Get model properties without compiling it.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
