% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxIndLin.R
\name{rxIndLinState}
\alias{rxIndLinState}
\title{Set the preferred factoring by state}
\usage{
rxIndLinState(preferred = NULL)
}
\arguments{
\item{preferred}{A list of each state's preferred factorization}
}
\value{
Nothing
}
\description{
Set the preferred factoring by state
}
\author{
Matthew Fidler
}
