% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RxODE.R
\name{rxInits}
\alias{rxInits}
\alias{rxInit}
\title{Initial Values and State values for a RxODE object}
\usage{
rxInits(obj, vec = NULL, req = NULL, defaultValue = 0,
  noerror = FALSE, noini = FALSE, rxLines = FALSE)

rxInit(obj, vec = NULL, req = NULL, defaultValue = 0,
  noerror = FALSE, noini = FALSE, rxLines = FALSE)
}
\arguments{
\item{obj}{rxDll, RxODE, or named vector representing default
initial arguments}

\item{vec}{If supplied, named vector for the model.}

\item{req}{Required names, and the required order for the ODE solver}

\item{defaultValue}{a number or NA representing the default value for
parameters missing in \code{vec}, but required in \code{req}.}

\item{noerror}{is a boolean specifying if an error should be thrown
for missing parameter values when \code{default} = \code{NA}}
}
\description{
Returns the initial values of the rxDll object
}
\author{
Matthew L.Fidler
}
\keyword{internal}
