% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxInv}
\alias{rxInv}
\title{Invert matrix using Rcpp Armadilo.}
\usage{
rxInv(matrix)
}
\arguments{
\item{matrix}{matrix to be inverted.}
}
\value{
inverse or pseudo inverse of matrix.
}
\description{
Invert matrix using Rcpp Armadilo.
}
