% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxModels_}
\alias{rxModels_}
\title{Get the rxModels  information}
\usage{
rxModels_(env = TRUE)
}
\arguments{
\item{env}{boolean that returns the environment where models are stored (TRUE), or the currently assigned RxODE model variables (FALSE).}
}
\description{
Get the rxModels  information
}
\keyword{internal}
