% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxNorm}
\alias{rxNorm}
\title{Get the normalized model}
\usage{
rxNorm(obj, condition = NULL, removeInis, removeJac, removeSens)
}
\arguments{
\item{obj}{RxODE family of objects}

\item{condition}{Character string of a logical condition to use
for subsetting the normalized model.  When missing, and a
condition is not set via \code{rxCondition}, return the whole
code with all the conditional settings intact.  When a
condition is set with \code{rxCondition}, use that condition.}

\item{removeInis}{A boolean indicating if parameter initialization
will be removed from the model}

\item{removeJac}{A boolean indicating if the Jacobians will be
removed.}

\item{removeSens}{A boolean indicating if the sensitivities will
be removed.}
}
\value{
Normalized Normal syntax (no comments)
}
\description{
This get the syntax preferred model for processing
}
\author{
Matthew L. Fidler
}
