% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxOptExpr.R
\name{rxOptExpr}
\alias{rxOptExpr}
\title{Optimize RxODE for computer evaluation}
\usage{
rxOptExpr(x)
}
\arguments{
\item{x}{RxODE model that can be access by rxNorm}
}
\value{
Optimized RxODE model text.  The order and type lhs and
    state variables is maintained while the evaluation is sped up.
    While parameters names are maintained, their order may be
    modified.
}
\description{
This optimizes RxODE code for computer evaluation by only
calculating redundant expressions once.
}
\author{
Matthew L. Fidler
}
