% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RxODE.R
\name{rxParams}
\alias{rxParams}
\alias{rxParam}
\title{Parameters specified by the model}
\usage{
rxParams(obj = NULL)

rxParam(obj = NULL)
}
\arguments{
\item{obj}{RxODE family of objects}
}
\value{
a character vector listing the parameters in the model.
}
\description{
This return the model's parameters that are required to solve the
ODE system.
}
\author{
Matthew L.Fidler
}
