% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxParams}
\alias{rxParams}
\alias{rxParam}
\title{Parameters specified by the model}
\usage{
rxParams(obj, constants = TRUE)

rxParam(obj, constants = TRUE)
}
\arguments{
\item{obj}{RxODE family of objects}

\item{constants}{is a boolean indicting if constants should be
included in the list of parameters. Currently RxODE parses
constants into variables in case you wish to change them
without recompiling the RxODE model.}
}
\value{
a character vector listing the parameters in the model.
}
\description{
This return the model's parameters that are required to solve the
ODE system.
}
\author{
Matthew L.Fidler
}
