% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxParseErr}
\alias{rxParseErr}
\title{Prepare Error function for inclusion in RxODE}
\usage{
rxParseErr(x, baseTheta, ret = "rx_r_", init = NULL)
}
\arguments{
\item{x}{error function}

\item{baseTheta}{Base theta to start numbering add(.) and prop(.) from.}

\item{ret}{Intenral return type.  Should not be changed by the user...}

\item{init}{Initilization vector}
}
\value{
RxODE transformed text
}
\description{
Prepare Error function for inclusion in RxODE
}
\author{
Matthew L. Fidler
}
\keyword{internal}
