% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxParseErr}
\alias{rxParseErr}
\title{Prepare Error function for inclusion in RxODE}
\usage{
rxParseErr(
  x,
  baseTheta,
  ret = "rx_r_",
  init = NULL,
  addProp = c("combined2", "combined1")
)
}
\arguments{
\item{x}{error function}

\item{baseTheta}{Base theta to start numbering add(.) and prop(.) from.}

\item{ret}{Internal return type.  Should not be changed by the user...}

\item{init}{Initialization vector}
}
\value{
RxODE transformed text
}
\description{
Prepare Error function for inclusion in RxODE
}
\author{
Matthew L. Fidler
}
\keyword{internal}
