% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxParsePk}
\alias{rxParsePk}
\title{Parse PK function for inclusion in RxODE}
\usage{
rxParsePk(x, init = NULL)
}
\arguments{
\item{x}{PK function}

\item{init}{Initialization vector}
}
\value{
RxODE transformed text.
}
\description{
Parse PK function for inclusion in RxODE
}
\author{
Matthew L. Fidler
}
\keyword{internal}
