% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode-options.R
\name{rxPermissive}
\alias{rxPermissive}
\alias{rxStrict}
\title{Permissive or Strict RxODE sytax options}
\usage{
rxPermissive(expr, silent = .isTestthat(), respect = FALSE,
  rxclean = .isTestthat(), cran = FALSE, on.validate = FALSE)

rxStrict(expr, silent = .isTestthat(), respect = FALSE,
  rxclean = .isTestthat(), cran = FALSE, on.validate = FALSE)
}
\arguments{
\item{expr}{Expression to evaluate in the permissive/strict
environment.  If unspecified, set the options for the current
environment.}

\item{silent}{when true, also silence the syntax errors and
interactive output (useful in testing).}

\item{respect}{when TRUE, respect any options that are specified.
This is called at startup, but really should not be called
elsewhere, otherwise the options are not changed.}

\item{rxclean}{when TRUE, call rxClean before and after the expr
is called.}

\item{cran}{When specifyed and true, run on CRAN. Otherwise it is skipped on cran.}

\item{on.validate}{When TRUE run only when validating.}
}
\description{
This sets the RxODE syntax to be permissive or strict
}
\author{
Matthew L. Fidler
}
