% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode-options.R
\name{rxPermissive}
\alias{rxPermissive}
\alias{rxStrict}
\title{Permissive or Strict RxODE syntax options}
\usage{
rxPermissive(
  expr,
  silent = .isTestthat(),
  respect = FALSE,
  cran = FALSE,
  on.validate = FALSE,
  test = NULL
)

rxStrict(
  expr,
  silent = .isTestthat(),
  respect = FALSE,
  cran = FALSE,
  on.validate = FALSE
)
}
\arguments{
\item{expr}{Expression to evaluate in the permissive/strict
environment.  If unspecified, set the options for the current
environment.}

\item{silent}{when true, also silence the syntax errors and
interactive output (useful in testing).}

\item{respect}{when TRUE, respect any options that are specified.
This is called at startup, but really should not be called
elsewhere, otherwise the options are not changed.}

\item{cran}{When specified and true, run on CRAN. Otherwise it is
skipped on CRAN.}

\item{on.validate}{When TRUE run only when validating.}

\item{test}{When specified as a string, the enclosed test is
skipped unless the environmental variable "rxTest" equals this
value.}
}
\description{
This sets the RxODE syntax to be permissive or strict
}
\author{
Matthew L. Fidler
}
