% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxPrune.R
\name{rxPrune}
\alias{rxPrune}
\title{Prune branches from RxODE}
\usage{
rxPrune(x)
}
\arguments{
\item{x}{RxODE model that can be accessed by rxNorm}
}
\value{
Pruned RxODE model text.  All logical expressions are
    collapsed to rx_lgl_# and are at the beginning of the file.
}
\description{
This prunse branches from RxODE.
}
\author{
Matthew L. Fidler
}
