% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxPrune.R
\name{rxPrune}
\alias{rxPrune}
\title{Prune branches (ie if/else) from RxODE}
\usage{
rxPrune(x)
}
\arguments{
\item{x}{RxODE model that can be accessed by rxNorm}
}
\value{
Pruned RxODE model text.
}
\description{
This prunes branches (ie if/else) from RxODE.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
