% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{rxRateDur}
\alias{rxRateDur}
\alias{as.rxRateDur}
\title{Creates a rxRateDur object}
\usage{
rxRateDur(x)

as.rxRateDur(x)
}
\arguments{
\item{x}{rxRateDur data}
}
\description{
This is primarily to display information about rate
}
