% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{rxRateDur}
\alias{rxRateDur}
\alias{[.rxRateDur}
\alias{as.rxRateDur}
\alias{as.character.rxRateDur}
\alias{[[.rxRateDur}
\alias{type_sum.rxRateDur}
\alias{pillar_shaft.rxRateDur}
\title{Creates a rxRateDur object}
\usage{
rxRateDur(x)

\method{[}{rxRateDur}(x, ...)

as.rxRateDur(x)

\method{as.character}{rxRateDur}(x, ...)

\method{[[}{rxRateDur}(x, ...)

type_sum.rxRateDur(x)

pillar_shaft.rxRateDur(x, ...)
}
\arguments{
\item{x}{rxRateDur data}

\item{...}{Other parameters}
}
\value{
rxRateDur object
}
\description{
This is primarily to display information about rate
}
