% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxReservedKeywords.R
\docType{data}
\name{rxReservedKeywords}
\alias{rxReservedKeywords}
\title{A list and description of Rode supported reserved keywords}
\format{
A data frame with 3 columns and 98 or more rows
\describe{
\item{Reserved Name}{Reserved Keyword Name}
\item{Meaning}{Reserved Keyword Meaning}
\item{Alias}{Keyword Alias}
}
}
\usage{
rxReservedKeywords
}
\description{
A list and description of Rode supported reserved keywords
}
\keyword{datasets}
