% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxS}
\alias{rxS}
\title{Load a model into a symengine environment}
\usage{
rxS(x, doConst = TRUE, promoteLinSens = FALSE)
}
\arguments{
\item{x}{RxODE object}

\item{doConst}{Load constants into the environment as well.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}
}
\value{
RxODE/symengine environment
}
\description{
Load a model into a symengine environment
}
\author{
Matthew Fidler
}
