% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxSetProd}
\alias{rxSetProd}
\title{Choose the type of product to use in RxODE.  These are used in the
RxODE \code{prod} blocks}
\usage{
rxSetProd(type = c("long double", "double", "logify"))
}
\arguments{
\item{type}{Product to use for \code{prod()} in RxODE blocks

\code{long double} converts to long double, performs the
multiplication and then converts back.

\code{double} uses the standard double scale for multiplication.}
}
\value{
nothing
}
\description{
Choose the type of product to use in RxODE.  These are used in the
RxODE \code{prod} blocks
}
\author{
Matthew L. Fidler
}
