% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rxSetProgressBar}
\alias{rxSetProgressBar}
\title{Set timing for progress bar}
\usage{
rxSetProgressBar(seconds = 1)
}
\arguments{
\item{seconds}{This sets the number of seconds that need to elapse
before drawing the next segment of the progress bar.  When
this is zero or below this turns off the progress bar.}
}
\description{
Set timing for progress bar
}
\author{
Matthew Fidler
}
