% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxShiny.R
\name{rxShiny}
\alias{rxShiny}
\alias{rxShiny.rxSolve}
\alias{rxShiny.default}
\title{Use Shiny to help develop an RxODE model}
\usage{
rxShiny(object, params = c(), events = NULL, inits = c(), ...)

\method{rxShiny}{rxSolve}(object, params = NULL, events = NULL,
  inits = c(), ...)

\method{rxShiny}{default}(object = NULL, params = c(), events = NULL,
  inits = c(), ...)
}
\arguments{
\item{object}{A RxODE family of objects. If not supplied a
2-compartment indirect effect model is used.  If it is
supplied, use the model associated with the RxODE object for
the model exploration.}

\item{params}{Initial parameters for model}

\item{events}{Event information (currently ignored)}

\item{inits}{Initial  estimates for model}

\item{...}{Other arguments passed to rxShiny.  Currently doesn't
do anything.}
}
\value{
Nothing; Starts a shiny server
}
\description{
Use Shiny to help develop an RxODE model
}
\author{
Zufar Mulyukov and Matthew L. Fidler
}
