% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{rxSolve}
\alias{rxSolve}
\title{Solves a rxDll object}
\usage{
rxSolve(object, params, events, inits = NULL, stiff = TRUE,
  transit_abs = FALSE, atol = 1e-08, rtol = 1e-06, ...)
}
\arguments{
\item{object}{is a RxODE family of objects.}

\item{params}{a numeric named vector with values for every
parameter in the ODE system; the names must correspond to the
parameter identifiers used in the ODE specification;}

\item{events}{an \code{EventTable} object describing the input
(e.g., doses) to the dynamic system and observation sampling
time points (see \code{\link{eventTable}});}

\item{inits}{a vector of initial values of the state variables
(e.g., amounts in each compartment), and the order in this
vector must be the same as the state variables (e.g., PK/PD
compartments);}

\item{stiff}{a logical (\code{TRUE} by default) indicating whether
    the ODE system is stifff or not.

    For stiff ODE sytems (\code{stiff=TRUE}), \code{RxODE} uses the
    LSODA (Livermore Solver for Ordinary Differential Equations)
    Fortran package, which implements an automatic method switching
    for stiff and non-stiff problems along the integration
    interval, authored by Hindmarsh and Petzold (2003).
      
    For non-stiff systems (\code{stiff=FALSE}), \code{RxODE} uses
    DOP853, an explicit Runge-Kutta method of order 8(5,3) of
    Dormand and Prince as implemented in C by Hairer and Wanner
    (1993).}

\item{transit_abs}{boolean indicating if this is a transit
compartment absorption}

\item{atol}{a numeric absolute tolerance (1e-08 by default);}

\item{rtol}{a numeric relative tolerance (1e-06 by default).}

\item{...}{Other arguments including scaling factors for each
compartment.  This includes S#=numeric will scale a compartment
# by a dividing the compartment amount by the scale factor,
like NONMEM.}
}
\value{
The output of \dQuote{rxSolve} is a matrix with as many
    rows as there are sampled time points and as many columns as
    system variables (as defined by the ODEs and additional
    assignments in the RxODE model code)
}
\description{
This uses a rxDll object to solve a system.
}
\references{
Hindmarsh, A. C.
\emph{ODEPACK, A Systematized Collection of ODE Solvers}.
Scientific Computing, R. S. Stepleman et al. (Eds.),
North-Holland, Amsterdam, 1983, pp. 55-64.

Petzold, L. R.
\emph{Automatic Selection of Methods for Solving Stiff and Nonstiff
Systems of Ordinary Differential Equations}.
Siam J. Sci. Stat. Comput. 4 (1983), pp. 136-148.

Hairer, E., Norsett, S. P., and Wanner, G.
\emph{Solving ordinary differential equations I, nonstiff problems}.
2nd edition, Springer Series in Computational Mathematics,
Springer-Verlag (1993).
}
\seealso{
\code{\link{RxODE}}
}

