% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{rxSumProdModel}
\alias{rxSumProdModel}
\title{Recast model in terms of sum/prod}
\usage{
rxSumProdModel(model, expand = FALSE, sum = FALSE, prod = TRUE,
  logify = FALSE)
}
\arguments{
\item{model}{RxODE model}

\item{expand}{Boolean indicating if the expression is expanded.}

\item{sum}{Use sum(...)}

\item{prod}{Use prod(...)}

\item{logify}{Use logify}
}
\value{
model string with prod(.) and sum(.) for all these
    operations.
}
\description{
Recast model in terms of sum/prod
}
\author{
Matthew L. Fidler
}
