% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxSymInv.R
\name{rxSymInvCholCreate}
\alias{rxSymInvCholCreate}
\title{Creates an object for calculating Omega/Omega^-1 and derivatives}
\usage{
rxSymInvCholCreate(mat, diag.xform = c("sqrt", "log", "identity"),
  create.env = TRUE, envir = parent.frame())
}
\arguments{
\item{mat}{Initial Omega matrix}

\item{diag.xform}{transformation to diagonal elements of OMEGA. or \code{chol(Omega^-1)}}

\item{create.env}{-- Create an environment to calculate the inverses. (By default TRUE)}

\item{envir}{-- Environment to evaluate function, bu default it is the parent frame.}
}
\value{
A rxSymInv object OR a rxSymInv environment
}
\description{
Creates an object for calculating Omega/Omega^-1 and derivatives
}
\author{
Matthew L. Fidler
}
\keyword{internal}
