% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyExists}
\alias{rxSymPyExists}
\title{Does the varaible exist in the SymPy Python environment?}
\usage{
rxSymPyExists(var)
}
\arguments{
\item{var}{Variable to be tested.}
}
\value{
boolean
}
\description{
Does the varaible exist in the SymPy Python environment?
}
\author{
Matthew L. Fidler
}
\keyword{internal}
