% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyFunctions}
\alias{rxSymPyFunctions}
\title{Setup SymPy functions}
\usage{
rxSymPyFunctions(functions)
}
\arguments{
\item{functions}{a list of functions}
}
\description{
This creates SymPy unspecified functions for later evaluation in the CAS SymPy.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
