% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetup}
\alias{rxSymPySetup}
\title{Setup a SymPy environment that sets up the specified RxODE model.}
\usage{
rxSymPySetup(model, envir = parent.frame())
}
\arguments{
\item{model}{RxODE family of objects}

\item{envir}{Environment to evaluate the model in (defaults to
parent frame)}
}
\value{
a boolean indicating if the environment is cached.
}
\description{
Setup a SymPy environment that sets up the specified RxODE model.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
