% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetupPred}
\alias{rxSymPySetupPred}
\title{Setup Pred function based on RxODE object.}
\usage{
rxSymPySetupPred(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL,
  grad = FALSE, logify = FALSE, pred.minus.dv = TRUE,
  theta.derivs = FALSE, grad.internal = FALSE, run.internal = FALSE,
  theta.internal = FALSE)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{grad}{Boolaen indicated if the the equations for the
gradient be calculated}

\item{logify}{Boolean to do multipication in log-space to reduce
round-off error.}

\item{pred.minus.dv}{Boolean stating if the FOCEi objective
function is based on PRED-DV (like NONMEM).  Default TRUE.}
}
\value{
RxODE object expanded with predfn and with calculated
    sensitivities.
}
\description{
This is for the so-called inner problem.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
