% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetupPred}
\alias{rxSymPySetupPred}
\title{Setup Pred function based on RxODE object}
\usage{
rxSymPySetupPred(obj, predfn, pkpars, errfn)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}
}
\value{
RxODE object expanded with predfn and with calculated sensitivities.
}
\description{
Setup Pred function based on RxODE object
}
\author{
Matthew L. Fidler
}
\keyword{internal}
