% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPySetupPred}
\alias{rxSymPySetupPred}
\title{Setup Pred function based on RxODE object.}
\usage{
rxSymPySetupPred(obj, predfn, pkpars = NULL, errfn = NULL, init = NULL,
  grad = FALSE, grad.internal = FALSE)
}
\arguments{
\item{obj}{RxODE object}

\item{predfn}{Prediction function}

\item{pkpars}{Pk Pars function}

\item{errfn}{Error function}

\item{init}{Initialization parameters for scaling.}

\item{grad}{Boolaen indicated if the the equations for the gradient be calculated}
}
\value{
RxODE object expanded with predfn and with calculated sensitivities.
}
\description{
This is for the so-called inner problem.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
