% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsympy.R
\name{rxSymPyVersion}
\alias{rxSymPyVersion}
\title{Return the version of SymPy that is running}
\usage{
rxSymPyVersion(numeric = TRUE)
}
\arguments{
\item{numeric}{boolean that specifies if the major and minor
release should be a number.}
}
\value{
Verson of sympy that is running.
}
\description{
Return the version of SymPy that is running
}
\author{
Matthew L. Fidler
}
