% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode-options.R
\name{rxSyncOptions}
\alias{rxSyncOptions}
\title{Sync options with RxODE variables}
\usage{
rxSyncOptions()
}
\description{
Accessing RxODE options via getOption slows down solving.  This
allows the options to be synced with variables.
}
\author{
Matthew L. Fidler
}
