% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxToSE}
\alias{rxToSE}
\alias{rxFromSE}
\alias{.rxFromSE}
\title{RxODE to symengine.R}
\usage{
rxToSE(x, envir = NULL, progress = FALSE, promoteLinSens = FALSE)

rxFromSE(x, unknownDerivatives = c("forward", "central", "error"))

.rxFromSE(x)
}
\arguments{
\item{x}{expression}

\item{envir}{default is \code{NULL}; Environment to put symengine
variables in.}

\item{progress}{shows progress bar if true.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}
}
\description{
RxODE to symengine.R
}
\author{
Matthew L. Fidler
}
