% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine.R
\name{rxToSE}
\alias{rxToSE}
\alias{rxFromSE}
\alias{.rxFromSE}
\title{RxODE to symengine.R}
\usage{
rxToSE(x, envir = NULL, progress = FALSE, promoteLinSens = TRUE)

rxFromSE(x, unknownDerivatives = c("forward", "central", "error"))

.rxFromSE(x)
}
\arguments{
\item{x}{expression}

\item{envir}{default is \code{NULL}; Environment to put symengine
variables in.}

\item{progress}{shows progress bar if true.}

\item{promoteLinSens}{Promote solved linear compartment systems to
sensitivity-based solutions.}

\item{unknownDerivatives}{When handling derivatives from unknown
    functions, the translator will translate into different types
    of numeric derivatives.  The currently supported methods are:

\itemize{

\item \code{forward} for forward differences

\item \code{central} for central differences

\item \code{error} for throwing an error for unknown derivatives

}}
}
\value{

}
\description{
RxODE to symengine.R
}
\author{
Matthew L. Fidler
}
