% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{rxToSymPy}
\alias{rxToSymPy}
\alias{rxFromSymPy}
\title{Converts model specification to/from a SymPy language}
\usage{
rxToSymPy(x, envir = parent.frame(1))

rxFromSymPy(x, envir = parent.frame(1))
}
\arguments{
\item{x}{is either RxODE family of objects, character of RxODE
expression or unquoted RxODE expression.}
}
\value{
Code Lines for sympy/RxODE that are named by what
    variables are defined.
}
\description{
Converts model specification to/from a SymPy language
}
\author{
Matthew L. Fidler
}
\keyword{internal}
