% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{rxUiGet}
\alias{rxUiGet}
\alias{rxUiGet.fun}
\alias{rxUiGet.ini}
\alias{rxUiGet.iniFun}
\alias{rxUiGet.modelFun}
\alias{rxUiGet.default}
\title{S3 for getting information from UI model}
\usage{
rxUiGet(x, ...)

\method{rxUiGet}{fun}(x, ...)

\method{rxUiGet}{ini}(x, ...)

\method{rxUiGet}{iniFun}(x, ...)

\method{rxUiGet}{modelFun}(x, ...)

\method{rxUiGet}{default}(x, ...)
}
\arguments{
\item{x}{list of (UIenvironment, exact).  UI environment is the parsed function for RxODE.  \code{exact} is a boolean that says}

\item{...}{Other arguments}
}
\value{
value that was requested from the UI object
}
\description{
S3 for getting information from UI model
}
\author{
Matthew Fidler
}
