% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rxUpdateResiduals}
\alias{rxUpdateResiduals}
\title{Update RxODE multi-subject data with new residuals (in-place).}
\usage{
rxUpdateResiduals(md)
}
\arguments{
\item{md}{The RxODE multi-data object setup from \code{\link{rxDataSetup}}}
}
\value{
A boolean indicating if this is a compatible object for updating residuals.
       If it isn't compatible nothing is done.  Additionally, if there are no random residual
       variables to update, also nothing is done.
}
\description{
Update RxODE multi-subject data with new residuals (in-place).
}
\author{
Matthew L. Fidler
}
\keyword{internal}
