% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\name{rxUse}
\alias{rxUse}
\title{Use model object in your package}
\usage{
rxUse(obj, internal = FALSE, overwrite = TRUE, compress = "bzip2")
}
\arguments{
\item{obj}{model to save.}
}
\description{
Use model object in your package
}
