% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxValidate.R
\name{rxValidate}
\alias{rxValidate}
\alias{rxTest}
\title{Validate RxODE}
\usage{
rxValidate(full = TRUE)

rxTest(full = TRUE)
}
\arguments{
\item{full}{Should a full validation be performed?  (By default
\code{TRUE})}
}
\description{
This allows easy validation/qualification of nlmixr by running the
testing suite on your system.
}
\author{
Matthew L. Fidler
}
