% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxValidate.R
\name{rxValidate}
\alias{rxValidate}
\alias{rxTest}
\title{Validate RxODE
This allows easy validation/qualification of nlmixr by running the
testing suite on your system.}
\usage{
rxValidate(type = NULL)

rxTest(type = NULL)
}
\arguments{
\item{type}{Type of test or filter of test type}
}
\value{
nothing
}
\description{
Validate RxODE
This allows easy validation/qualification of nlmixr by running the
testing suite on your system.
}
\author{
Matthew L. Fidler
}
