% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{rxVersion}
\alias{rxVersion}
\title{Version and repository for this dparser package.}
\usage{
rxVersion(
  extra = "",
  echo = FALSE,
  version = sessionInfo()$otherPkgs$RxODE$Version
)
}
\arguments{
\item{extra}{extra text to display in the logo}

\item{echo}{Boolean to echo the text logo, by default FALSE}

\item{version}{Version to display/return}
}
\value{
A character vector with the version and repository.
}
\description{
Version and repository for this dparser package.
}
\author{
Matthew L. Fidler
}
\keyword{internal}
