% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE-win-setup.R
\name{rxWinSetup}
\alias{rxWinSetup}
\title{Setup Windows components for RxODE}
\usage{
rxWinSetup(rm.rtools = TRUE, rm.python = TRUE)
}
\arguments{
\item{rm.rtools}{Remove the Rtools from the current path specs.}

\item{rm.python}{Remove Python from the current path specs.}
}
\description{
Setup Windows components for RxODE
}
\author{
Matthew L. Fidler
}
