% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxsolve-gen.R
\name{sample_n}
\alias{sample_n}
\alias{sample_n.solveRxDll}
\title{sample_n for \code{solveRxDll} object}
\usage{
\method{sample_n}{solveRxDll}(tbl, ...)
}
\arguments{
\item{tbl}{Solved equation, an \code{solveRxDll} object.}

\item{...}{Additional arguments}
}
\description{
compatability function for dplyr
}
