% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{solve.RxODE}
\alias{solve.RxODE}
\title{Solve RxODE objects}
\usage{
\method{solve}{RxODE}(...)
}
\arguments{
\item{...}{Additional arguments sent to \code{rxSolve}}
}
\description{
Solve RxODE objects
}
\seealso{
\code{\link{rxSolve}}
}
\author{
Matthew L.Fidler
}
