% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{solve.RxODE}
\alias{predict.RxODE}
\alias{solve.RxODE}
\title{Solve an RxODE object}
\usage{
\method{solve}{RxODE}(RxODEobj, ...)

\method{predict}{RxODE}(RxODEobj, ...)
}
\arguments{
\item{RxODEobj}{An rxode object}

\item{...}{Solve arguments sent to \code{RxODEobj$solve}.  See
\code{\link{RxODE}}.}
}
\description{
\code{solve} solves the odinary differential equations specified by
 RxODE object
}

