% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{solve.RxODE}
\alias{solve.RxCompilationManager}
\alias{solve.RxODE}
\alias{solve.character}
\alias{solve.rxDll}
\alias{solve.solveRxDll}
\title{Solve RxODE objects}
\usage{
\method{solve}{RxODE}(...)

\method{solve}{RxCompilationManager}(...)

\method{solve}{solveRxDll}(...)

\method{solve}{character}(...)

\method{solve}{rxDll}(...)
}
\arguments{
\item{...}{Additional arguments sent to \code{rxSolve}}
}
\description{
Solve RxODE objects
}
\seealso{
\code{\link{rxSolve}}
}

