% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rxDll}
\alias{summary.rxDll}
\title{Summary of rxDll object}
\usage{
\method{summary}{rxDll}(object, ...)
}
\arguments{
\item{object}{RxDll object}

\item{...}{Other arguments.  Includes \code{noprint}, which is a
logical telling if the object should print the rxDll object
first. By default this is FALSE}
}
\value{
object is returned
}
\description{
This gives expanded information about the rxDll object
}
\author{
Matthew L.Fidler
}
\keyword{internal}
