% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RxODE.R
\name{summary.RxODE}
\alias{summary.RxCompilationManager}
\alias{summary.RxODE}
\title{Print expanded information about the RxODE object.}
\usage{
\method{summary}{RxODE}(object, ...)

\method{summary}{RxCompilationManager}(object, ...)
}
\arguments{
\item{object}{RxODE object}

\item{...}{Ignored parameters}
}
\description{
This prints the expanded information about the RxODE object.
}

