% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.RxODE}
\alias{summary.RxODE}
\title{Print expanded information about the RxODE object.}
\usage{
\method{summary}{RxODE}(object, ...)
}
\arguments{
\item{object}{RxODE object}

\item{...}{Ignored parameters}
}
\value{
object is returned
}
\description{
This prints the expanded information about the RxODE object.
}
\author{
Matthew L.Fidler
}
